% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRecord.R
\name{addRecord}
\alias{addRecord}
\title{Adds a record to a table along with an (optional) associated file}
\usage{
addRecord(con, table_name, new_record, add_file = FALSE, 
storage_type = NA,  file_path = NA)
}
\arguments{
\item{con}{An open connection to the database}

\item{table_name}{A string, the name of the table to add the record to}

\item{new_record}{A dataframe of new records, with 
columns that match the table to be appended to.}

\item{add_file}{TRUE/FALSE.  Whether to add an associated file, if any 
(default FALSE)}

\item{storage_type}{Where files associated with new records should be stored.
Options are "local", "Google Drive", "AWS S3", or "SharePoint". If NA, the
files are assumed to be stored in subdirectories of an AMMonitor project.}

\item{file_path}{Filepath of file associated with new db record}
}
\value{
Returns a list with the status of the addition (TRUE if successful) and
a message indicating any errors encountered, if the addition failed.
}
\description{
Adds a record to a table along with an (optional) associated file
in local, Google Drive, AWS S3, or SharePoint directories.
}
\details{
When adding a record to the AMMonitor database, the media, models, 
logs, and spatials tables all reference external files, which are tracked with 
the  filepath field (media, spatials, and logs tables) or amml field (models 
table). The addRecord() function performs a regular insert query for all other 
tables, but for these four, this function attempts to add the associated file 
before adding the database record, when specified, based on the associated 
filepath. This function performs a copy of the specified file to the location
specified in the new database record's filepath/amml column.

See the "database" learnr tutorial for more details on the AMMonitor database
tables. The tutorial can be launched with \code{learnr::run_tutorial(name = 
"database", package = "AMMonitor")}. For more details on cloud storage options
see the "media" tutorial.
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in a temporary directory
# (to be deleted)

# run the function and capture the connection
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# add a person to the database
new_person <- data.frame(pk_personid = 'test_person', display_name = 'test')

rs <- addRecord(con = conx, table_name = 'people', new_record = new_person)

# print the status of the addition
print(rs)

# demonstrate an error returned when addRecord fails
rs <- addRecord(con = conx, table_name = 'people', new_record = new_person)
  
# print the status of the addition
print(rs)

# Demonstrate adding a record with a file ------------------

# photo to add to the database (use one of the demo photos)
file_source <- dir(
  file.path(
    find.package("AMMonitor", lib.loc = .libPaths()), 
    "extdata/demoAMM/photos"
  ), 
  full.names = T
)[1]

# where the new photo should be saved (AMMonitor project's "photos" directory)
file_destination <- file.path(demo_fp, 'photos/test.JPG')

# create dataframe to be sent to the addRecord function
new_media <- data.frame(
  media_type = "photo",
  filename = "test.JPG",
  fk_visitid = 1,
  start_date = "1900-01-01",
  start_time = "00:00:00",
  filepath = file_destination
)

# add the record with media
rs <- addRecord(
  con = conx,
  table_name = "media",
  new_record = new_media,
  add_file = TRUE,
  storage_type = "local",
  file_path = file_source
)

# print the status of the addition
print(rs)

# confirm the new file was added
file.exists(new_media$filepath)

# disconnect from the database when finished
DBI::dbDisconnect(conx)

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)

}
}
\seealso{
Other database: 
\code{\link{dbAddUserCol}()},
\code{\link{dbAddUserTable}()},
\code{\link{dbCheckCore}()},
\code{\link{dbCheckData}()},
\code{\link{dbCheckup}()},
\code{\link{dbCreate}()},
\code{\link{dbImportCSV}()},
\code{\link{dbUpdateVersion}()},
\code{\link{deleteRecord}()}
}
\concept{database}
