% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammCreateDirectories.R
\name{ammCreateDirectories}
\alias{ammCreateDirectories}
\title{Create file structure for an AMMonitor program}
\usage{
ammCreateDirectories(amm_dirname, filepath)
}
\arguments{
\item{amm_dirname}{The name of the directory that will house an AMMonitor 
project.}

\item{filepath}{File path to the directory that will house the database.}
}
\value{
The filepath of the AMMonitor project.
}
\description{
Create the structure required for an AMMonitor project. This is 
the first function users invoke when creating an AMMonitor project. Note 
that some AMMonitor functions rely on this specific directory structure 
which we assume users will implement. In particular, the SQLite database
should always be stored in the "database" directory, and default "settings"
should always be stored in the "settings" directory.
}
\details{
See the "getting_started" tutorial for more information.  Each empty 
directory generated by this function will store specific types of information, 
as detailed below:
\itemize{
  \item \strong{ammls}: Stores AMModel libraries
  \item \strong{database}: Stores the SQLite database
  \item \strong{logs} and \strong{log_drop}: Stores archived logs that 
  provide information about
  an analysis or an equipment performance
  \item \strong{ml_drop}: Stores machine learning outputs ready for import in
  the database.  
  \item \strong{mobile_apps}: Stores templates for collecting field-based
  information via mobile applications such  as Survey123 or Google
  AppSheets.
  \item \strong{photos} and \strong{photos_drop}: Stores archived photos 
  \item \strong{recordings} and \strong{recordings_drop}: Stores archived 
  audio recording files (e.g., .wav) 
  captured in acoustic monitoring programs
  \item \strong{scripts}: Stores R scripts that can be sourced  to automatically 
  process new data
  \item \strong{settings}: Stores files needed to access accounts through R
  \item \strong{spatials}: Stores spatial layers associated with locations 
  in a monitoring program (rasters and/or shapefiles) as RDS files
  \item \strong{videos} and \strong{video_drop}: Stores archived videos 
}

The "drop" folders can be used to temporarily store media files or log files
until they are processed and entered into the database, and then moved into 
a more permanent directory.   

It is possible to store media files outside the AMMonitor project directory,
but the tutorials highlight the default file structure.
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in your working directory
# (to be deleted)
fp <- ammCreateDirectories(amm_dirname = "myProject", filepath = tempdir())

# look at the filepath of the AMMonitor project
fp 

# look at the folders within an AMMonitor project
list.files(fp, recursive = FALSE)

# remove the demo AMMonitor file structure
unlink(fp, recursive = TRUE)
}
}
