% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammCreateMiniDemo.R
\name{ammCreateMiniDemo}
\alias{ammCreateMiniDemo}
\title{Create a lightweight AMMonitor project for demonstration}
\usage{
ammCreateMiniDemo(filepath = tempdir())
}
\arguments{
\item{filepath}{File path to the directory that will house the database.
Default is tempdir().}
}
\value{
Returns the filepath of the mini AMMonitor project.
}
\description{
Create a lightweight AMMonitor project for demonstration. An
AMMonitor project consists of several directories and an AMMonitor SQLite 
database that tracks all components of a monitoring program.
}
\details{
Each 
AMMonitor project consists of a set of directories that store specific types 
of information, as detailed below:
\itemize{
  \item \strong{ammls}: Stores AMModel libraries.  A demo library is named
  "templates.RDS", which stores 4 templates created with the monitoR package.
  \item \strong{database}: Stores the SQLite database. The example database
  is named "demo.sqlite".
  \item \strong{logs} and \strong{log_drop}: Stores archived logs that 
  provide information about
  an analysis or an equipment performance.  
  \item \strong{videos} and \strong{video_drop}: Stores archived videos. 
  (Video functions are yet to be developed). 
  \item \strong{photos} and \strong{photos_drop}: Stores archived photos. The
  example photos directory consists of 25 photos.
  \item \strong{recordings} and \strong{recordings_drop}: Stores archived 
  audio recording files (e.g., .wav) 
  captured in acoustic monitoring programs.  The example recordings directory
  stores one recording named "1.wav".
  \item \strong{scripts}: Stores R scripts that can be sourced to automatically 
  process new data.  The example script is named "dbCheckAll.R".
  \item \strong{settings}: Stores files needed to access accounts through R.
  The demo settings include "default_user.txt",  "audio_path.txt",
   "image_path.txt", and  "noaa_token.txt".
  \item \strong{spatials}: Stores spatial layers associated with locations 
  in a monitoring program (rasters and/or shapefiles) as raw files or RDS files.
}


See the "getting_started" learnr tutorial for more details on what an 
AMMonitor project is and its setup.  
The tutorial can be launched with \code{learnr::run_tutorial(name = 
"getting_started", package = "AMMonitor")}.
}
\examples{
\dontrun{

# create a demo AMMonitor project in a temporary directory
# (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# look at the folders within an AMMonitor project
list.files(demo_fp, recursive = FALSE)

# look at the sample photos in the photos directory
list.files(file.path(demo_fp, "photos"), recursive = FALSE)

# look at the sample recording in the recordings directory
list.files(file.path(demo_fp, "recordings"), recursive = FALSE)

# look at the sample ammodels library that stores templates in the amml directory
list.files(file.path(demo_fp, "ammls"), recursive = FALSE)

# read in the "templates" model library; note there are 4 templates
templates <- readRDS(file.path(demo_fp, "ammls", "templates.RDS"))
templates

# look at the sample settings in the settings directory
list.files(file.path(demo_fp, "settings"))

# see who the default user is for the database
readLines(file.path(demo_fp, "settings", "default_user.txt"))

# look at the sample scripts in the scripts directory
list.files(file.path(demo_fp, "scripts"))

# look at the sample survey123 template in the mobile_apps directory
list.files(file.path(demo_fp, "mobile_apps"))

# look at the sample database in the database directory
list.files(file.path(demo_fp, "database"))

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# list the tables in the AMMonitor database
DBI::dbListTables(conx)

# disconnect from the database when finished
DBI::dbDisconnect(conx)

# to work with the shiny app, just point to the AMMonitor project directory
# the app will set the database connection and disconnect
launchApp(amm_project_path = demo_fp)

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)

}
}
