% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMMonitor-package.R
\docType{package}
\name{AMMonitor-package}
\alias{AMMonitor}
\alias{AMMonitor-package}
\title{AMMonitor: Remote monitoring of biodiversity in an adaptive framework}
\description{
AMMonitor is a multi-purpose biodiversity/species monitoring platform that utilizes (1) Autonomous Monitoring Units (AMUs) such as trail cameras, ARUs or smartphones to collect data, (2) SQLite as a database engine for storing and tracking all components of the monitoring effort, and (3) a suite of R-based functions for analysis of monitoring data. AMMonitor is a highly extensible monitoring system that can be adapted to multiple analytical frameworks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://code.usgs.gov/vtcfwru/ammonitor}
  \item Report bugs at \url{https://code.usgs.gov/vtcfwru/ammonitor/issues}
}

}
\author{
\strong{Maintainer}: Laurence Clarfeld \email{laurence.clarfeld@uvm.edu}

Authors:
\itemize{
  \item Caroline Tang \email{caroline.tang77@gmail.com}
  \item Kaitlin Huber \email{kehuber97@gmail.com}
  \item Cathleen Balantic \email{cbalantic@gmail.com}
  \item Therese Donovan \email{tdonovan@uvm.edu}
}

Other contributors:
\itemize{
  \item Kayley Dillon \email{kayley.dillon@uvm.edu} [contributor]
}

}
\keyword{internal}
