% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierPredict.R
\name{classifierPredict}
\alias{classifierPredict}
\title{Classify template detections into true and false positives}
\usage{
classifierPredict(con,
 classifierName,
 recordingNames = 'all',
 recordingRootPath = 'recordings/',
 ammlPath = 'ammls/',
 dbInsert = FALSE,
 showProgress = FALSE)
}
\arguments{
\item{con}{An open database connection}

\item{classifierName}{A character vector specifying which classifier model to 
run.}

\item{recordingNames}{Default = 'all'. An optional character vector specifying 
by name which recordings should be processed.}

\item{recordingRootPath}{A character string indicating the path to the 
directory 
containing the recordings, can be local or a URL. If NULL, will use the path 
indicated by the media.filepath column in the database.}

\item{ammlPath}{A character string indicating the path to the directory where
AMModel Libraries are stored.}

\item{dbInsert}{Default = FALSE. Logical flag for whether to insert records 
generated by this function into the database. If TRUE, events are added to the 
database. If FALSE, events are returned as a data.table for examination by 
the user, but not added to the database.}

\item{showProgress}{Default = FALSE. If TRUE, progress will be reported during 
the retrieval of data.}
}
\value{
Data.table of scores
}
\description{
\code{classifierPredict} runs classifier models (which are 
trained 
to classify the detections from a particular template) on all detections of 
a 
template in order to determine whether the detections are true or false 
positives. Please see the "models" tutorial for information on this function.
\code{learnr::run_tutorial("models", package = "AMMonitor")}
}
