% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbAddUserTable.R
\name{dbAddUserTable}
\alias{dbAddUserTable}
\title{Add a user table to an AMMonitor database}
\usage{
dbAddUserTable(con, table_name, table_info = NULL, 
excel_fp = NA, primary_tab = NA, subtable = NA, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{table_name}{The name of the new table to be added}

\item{table_info}{A dataframe with new table information. Use this option only
if your new table does not contain columns that reference lists. If 
your table does contain a column referencing a list, enter the information 
based on the Excel template (See Details).}

\item{excel_fp}{Filepath to an Excel file with the new table information}

\item{primary_tab}{Primary tab under which the table should appear in shiny}

\item{subtable}{Table under which records should appear as a subtable in shiny}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE}
}
\value{
Returns 1 if successful
}
\description{
Add a unique user-defined table to an AMMonitor database and 
update the dbdictionary and shinytables table to reflect changes in the 
database schema.
}
\details{
Adding new tables to the AMMonitor database requires a detailed 
knowledge of database structure and design. Any added user tables will be 
supported by the Shiny app's database front-end. New tables can be specified 
by completing the template Excel file that comes with AMMonitor and is named
"dbUserTable.xlsx".  This template can be copied to your working directory 
with the following code:

 \code{fp <- system.file("extdata/dbUserTable.xlsx", package = "AMMonitor")}
 
 \code{file.copy(from = fp, to = getwd(), overwrite = FALSE)}
 
 From there, open the Excel file and fill it in. Note that the fields here come 
 from the dbdictionary table, and review of both the "database" and "dbdictionary"
 tutorials is recommended before adding a user table. 
 
 The template contains a sample user table named "vegsamples" as an example, 
 which is a table for storing vegetation sampling results for a given location.
 
 Once the Excel file is filled out with new table information, point to this 
 new file to run the dbAddUserTable() function. 
 
 The Excel file has three sheets:
 
 Sheet 1: dbUserTableInfo:
  Columns in this sheet are defined in the dbdictionary table and can be returned  
  with the following query after a connection to the database is made:
  
 \code{DBI::dbListFields(conx, table = "dbdictionary")}

Sheet 2: lists:  
 If the new database user table will contain a column with entries limited 
to a list, create the list name on this sheet.  Use lower case for list 
names.

Sheet 3: listitems: 
 If the new database user table will contain a column with entries limited 
to a list, create the acceptable list entries on this sheet. You must identify
the list name (fk_listid), the item, and a description of the item.

Table info can be passed to the \code{dbAddUserTable()} function
excel template or as a data.frame. Additionally, the function will accept a
"primary_tab" argument that defines which tab the table will appear under in 
the AMMonitor Shiny app's database front-end, as well as a subtable to include 
the new user table under, if applicable.

See the "database" and "dbdictionary" learnr tutorials for more details on the
database. The tutorials can be launched with:

\code{learnr::run_tutorial(name = "database", package = "AMMonitor")}.
\code{learnr::run_tutorial(name = "dbdictionary", package = "AMMonitor")}.
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in a temporary directory
# (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database/demo.sqlite"))

# look at the tables in the demo database (these are AMMonitor core-tables)
DBI::dbListTables(conx)

# New table name
table_name <- "grants"

# Set the primary tab under which the new table will appear in the Shiny app
primary_tab <- "Program Mgt"

# Create the new "grants" table; does not have any columns with dropdowns
new_table <-  data.frame(
  pk_fieldname = c("pk_grantid", "grantor", "award_date", "total_award", "notes"),
  var_type = c("INTEGER", "VARCHAR(255)", "VARCHAR(255)", "REAL", "TEXT"),
  not_null_clause = NA,
  pk = c(1, 0, 0, 0, 0),
  foreign_key_table = NA,
  foreign_key_field = NA,
  on_update = NA,
  on_delete = NA,
  shiny_placeholder = NA,
  shiny_input = c("locked", "text", "date", "numeric", "longtext"),
  default_value = NA,
  fk_listid = NA,
  sb_include = 0,
  min = NA,
  max = NA,
  description = c(
    "Grant ID (auto-number)",
    "Granting agency (i.e., NSF, NIH, etc.)",
    "Date the grant was awarded",
    "Value of grant (in dollars)",
    "Notes about the grant"
    ),
 sort_order = 1:5
 )
 
# add the new table to the database
AMMonitor::dbAddUserTable(
 con = conx, 
 table_name = table_name,
 table_info = new_table,
 primary_tab = primary_tab,
 disconnect = FALSE
 )
 
 # confirm the new table is added
 DBI::dbListTables(conx)
 DBI::dbListFields(conx, name = "grants")
 
  
 # check the dbdictionary table for the new table
 DBI::dbGetQuery(
  con = conx,
  statement = "SELECT * from dbdictionary WHERE pk_tablename = 'grants';")
  
 # double check using SQLite table information
 DBI::dbGetQuery(
  con = conx,
  statement = "PRAGMA table_info(grants)"
  )
 
 # add a record to the new table
 DBI::dbAppendTable(conx,
    name = "grants",
    value = data.frame(
      pk_grantid = 1,
      grantor = "Agency XYZ",
      award_date = "2022-01-01",
      total_award = 10000,
      notes = "My award notes.")
  )
  
 # confirm the record
 DBI::dbReadTable(conx, name = "grants")
 
 # look at the table through the shiny app (click on database tab)
 launchApp(demo_fp)

# ------------------------------------------------------------
# if user tables include columns with a dropdown, use the Excel template.
# Add the example user table that comes with the AMMonitor package
# Here, we will be adding a user table named "vegsamples" to
# the database, containing a foreign key reference to the "locations" table
# and several columns with possible values restricted to a list.
# ------------------------------------------------------------


# Set the primary tab under which the new table will appear in the Shiny app
primary_tab <- "Location Info"

# set the filepath to the Excel file that has the required information
fp <- system.file("extdata/dbUserTable.xlsx", package = "AMMonitor")

# glimpse (but don't load) at sheets in the dbUserTable spreadsheet 
readxl::read_excel(
 path = fp,
 sheet = "dbUserTableInfo", 
 col_types = c(
 "text", "text", "text", "numeric", "text", "text", "text", "text", "text", 
 "text", "text", "text", "numeric", "numeric", "numeric", "numeric", "text")
 )
 
# glimpse at the lists
readxl::read_excel(path= fp, sheet = "lists")
  
# glimpse at the listitems
readxl::read_excel(path = fp, sheet = "listitems")

# add the template user table to the database
dbAddUserTable(
  con = conx,
  table_name = "vegsamples",
  excel_fp = fp,
  primary_tab = primary_tab,
  disconnect = FALSE)

# list the fields in the new tables and note the new user columns
DBI::dbListFields(conx, name = "vegsamples")

# disconnect from the database when finished
DBI::dbDisconnect(conx)

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)


}
}
\seealso{
\code{\link{dbCreate}}, \code{\link{dbAddUserCol}}

Other database: 
\code{\link{addRecord}()},
\code{\link{dbAddUserCol}()},
\code{\link{dbCheckCore}()},
\code{\link{dbCheckData}()},
\code{\link{dbCheckup}()},
\code{\link{dbCreate}()},
\code{\link{dbImportCSV}()},
\code{\link{dbUpdateVersion}()},
\code{\link{deleteRecord}()}
}
\concept{database}
