% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCheckCore.R
\name{dbCheckCore}
\alias{dbCheckCore}
\title{Compares the core fields and lists in an AMMonitor  database with 
the built-in AMMonitor default dictionary}
\usage{
dbCheckCore(con, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to a database}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE}
}
\description{
Checks a database to find mismatches between it and the core columns of the default AMMonitor database
}
\details{
This function should be used occasionally to ensure that the a 
given AMMonitor database that is in use has the "core" tables, "core" fields,
and "core" lists that come with a new, default AMMonitor database.  Most 
AMMonitor functions rely on this "core" setup; this function checks for 
any inconsistencies and provides output that points to them.  This function
will NOT fix the inconsistencies; rather it is up to the user to fix them
in their database.

The output is a list with seven sections. One is a list "description" that 
briefly explains errors that were found. Each of the other six are 
data.frames containing mappings to any mismatches between the input database 
and the default AMMonitor 
database. The "dictionary_check" data.frame shows discrepancies between the 
database and the default within core fields of the "dbdictionary" table. 
Similarly, each of the other data.frames returned in the output will reveal 
mismatches for the "lists" and "listitems" tables, the "librarylists" table, 
the "librarylistitems" table, the "medialists" table, and the 
"medialistitems" table. In each of these data.frames, the prefix "db_" 
describes information from the database passed into the dbCheckCore 
function, and the prefix "default_" describes information pertaining to the 
default AMMonitor database. 

#' See the "database" learnr tutorial for more details on the
database. The tutorial can be launched with
\code{learnr::run_tutorial(name = "database", package = "AMMonitor")}.
}
\examples{
\dontrun{

# create a demo AMMonitor project in a temporary directory
# (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# list the tables in the AMMonitor database
DBI::dbListTables(conx)

# check the database core columns against the AMMonitor default
results <- dbCheckCore(con = conx, disconnect = FALSE)
 
# the errors, if present, will be given in the captured output
str(results)

# -----------------------------------------------------
# the demo has no errors
# let's change the demo dictionary so you can see the errors returned

# adding dictionary mismatch
DBI::dbExecute(
 conx, 
 statement = "UPDATE dbdictionary SET pk_fieldname = 'primary' 
   WHERE pk_fieldname = 'primary_account';"
 )

# adding librarylist mismatch
DBI::dbExecute(
 conx, 
 statement = "UPDATE librarylists 
   SET pk_librarylistid = 'ageClass' 
   WHERE pk_librarylistid = 'age_class';"
 )


# check the database core columns against the AMMonitor default
results <- dbCheckCore(con = conx, disconnect = TRUE)

# look at results
lapply(results,  FUN = head)

# -----------------------------------------------------

# remove the demo AMMonitor project
unlink(demo_fp, recursive = TRUE)

}
}
\seealso{
Other database: 
\code{\link{addRecord}()},
\code{\link{dbAddUserCol}()},
\code{\link{dbAddUserTable}()},
\code{\link{dbCheckData}()},
\code{\link{dbCheckup}()},
\code{\link{dbCreate}()},
\code{\link{dbImportCSV}()},
\code{\link{dbUpdateVersion}()},
\code{\link{deleteRecord}()}
}
\concept{database}
