% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCheckData.R
\name{dbCheckData}
\alias{dbCheckData}
\title{Performs basic database integrity checks}
\usage{
dbCheckData(con, check_list = NA, verbose = TRUE, disconnect = FALSE, ...)
}
\arguments{
\item{con}{An open database connection}

\item{check_list}{List of data quality checks to perform (default NA for "all")}

\item{verbose}{Whether or not to print the status of each check (default TRUE)}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE}

\item{...}{Optional arguments to be used in checks (see Details)}
}
\description{
Checks database integrity by checking for inconsistencies in
data entries.
}
\details{
This function performs data checks to ensure data is in the right format and
is logically coherent. A data.frame is returned that names each data issue
encountered, with a description and advice for resolving the issue.
By default, all checks are performed.

Checks that can be performed by this function include:
\itemize{
\item visit_date: all visit dates are in a valid format
\item visit_time: all visit times are in a valid format
\item media_time: all media dates are in a valid format
\item media_time: all media times are in a valid format
\item visit_type_sequence: all visits appropriately sequenced by type
\item media_start_times: media dates/times consistent with their associated visits
}

Note that the natural order that checks are run is deliberate, and that it is
sometimes the case that the first issues presented must be addressed before
later issues. For example, the media_start_times check may require that all
visit and media dates/times are in a valid format to work properly.

For the media_start_times check, the optional parameter "time_threshold" can
be entered to specify a buffer time (in seconds) around the visit time to
allow for media times to fall outside of the specified visits by a given amount
without returning a warning. The default for this parameter is 60\*60\*6 =21600
seconds, or 6 hours.

Some other helpful AMMonitor functions for checking data quality are:
\itemize{
\item dbCheckup: Checks for foreign key constraint violations and mismatches
between the dbDictionary and database schema.
\item dbGetSummaryData: Creates a list of dataframes from a database for summary
purposes. This list of dataframes can then be used to create a variety of
summary statistics and figures using \code{dbPlotSummary()} or
\code{dbTableSummary()}, which can be helpful for spotting outliers that
may result from data entry mistakes.
}

See the "database" learnr tutorial for more details on the
database. The tutorial can be launched with
\code{learnr::run_tutorial(name = "database", package = "AMMonitor")}.
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in a temporary directory
# (to be deleted)

# run the function and capture the connection
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# list the tables in the AMMonitor database
DBI::dbListTables(conx)

# look at the media table
DBI::dbReadTable(conx, name = "visits")

# check the database; do not disconnect on exit
results <- dbCheckData(con = conx, disconnect = FALSE)
 
# the errors, if present, will be given in the captured output
str(results)

# -----------------------------------------------------
# the demo has no errors
# let's change some data entries so you can see the errors are returned

# introduce an out-of-sequence pull visit before a new set at this location
DBI::dbAppendTable(
    conx,
    name = 'visits',
    value = data.frame(
      pk_visitid = 100,
      fk_personid = "fbaggins",
      fk_locationid = "locationA",
      fk_equipmentid = "camera1",
      visit_type = "pull", 
      visit_date = "2024-01-01",
      visit_time = "12:00:00",
      visit_notes = NA,
      fk_econfigid = NA)
 )

# add a visit with invalid visit time
DBI::dbAppendTable(
    conx,
    name = 'visits',
    value = data.frame(
      pk_visitid = 101,
      fk_personid = "fbaggins",
      fk_locationid = "locationB",
      fk_equipmentid = "camera2",
      visit_type = "set", 
      visit_date = "2024-01-01",
      visit_time = "2:00",
      visit_notes = NA,
      fk_econfigid = NA
    ))
 
# run the check again, disconnect on exit
results <- dbCheckData(con = conx, disconnect = TRUE)

# look at the results
str(results)

# -----------------------------------------------------

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)

}
}
\seealso{
Other database: 
\code{\link{addRecord}()},
\code{\link{dbAddUserCol}()},
\code{\link{dbAddUserTable}()},
\code{\link{dbCheckCore}()},
\code{\link{dbCheckup}()},
\code{\link{dbCreate}()},
\code{\link{dbImportCSV}()},
\code{\link{dbUpdateVersion}()},
\code{\link{deleteRecord}()}
}
\concept{database}
