% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCheckUp.R
\name{dbCheckup}
\alias{dbCheckup}
\title{Performs basic database integrity checks}
\usage{
dbCheckup(con, errors_only = TRUE, disconnect = FALSE)
}
\arguments{
\item{con}{An open database connection}

\item{errors_only}{TRUE or FALSE.  Should only database errors be returned?}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE}
}
\description{
Checks database integrity by identifying any foreign key
constraints, as well as checking if the dbdictionary table matches
the current database schema.
}
\details{
dbCheckup() is an important function to run occasionally to ensure
that there are no foreign key violations present in the database, as well
as to ensure that the dbdictionary database table accurately reflects the
current database schema (tables and columns).  This is especially important
when releasing data to the public on ScienceBase, as the release will include
the dbdictionary table as a csv file and is used to re-create the original
release data with information on the ScienceBase release page.

The output is a list with four sections- one a list "description" with a 
short explanation of any errors found, and three data.frames containing
any errors encountered during the checks.  The 
goal is to yield empty data.frames in each section.  In "violations", a 
data.frame is returned that provides a list of foreign key violations.  
The offending table and row is provided in the columns "table" and "rowid".  
The parent table that enforces the keys is given, as well as the foreign key
identification "fkid". A violation means that an entry in 
the table that has a foreign key restriction contains data that violates the
key.

The second section named "tables" provides a check of the database structure.
It compares the actual database structure defined in its schema (column name,
variable type) with those identified in the database dictionary table.  
Columns with the prefix "db_" indicate information generated by the actual
database, while the prefix "dict_" indicates what is stored in the 
dictionary. The column "col_test" tests if the columns names match; the 
column "pk_test" tests if the database primary keys match the dictionary;
the column "type_test" tests if the database data type matches the
dictionary. The column "errors" flags whether any mismatches occur.

The final list element is "foreign_keys", and this data.frame returns a 
check between the foreign key constraints as specified in the database 
schema with those identified in the dbdictionary table.
The columns "missing_key_test", "child_test", and "parent_test", 
"on_update_test", "on_delete_test" should be TRUE if they pass. The 
"missing_key_test" checks that the key exists. The "child_test" checks that 
the database fieldname matches the dbdictionary fieldname for that key. The 
"parent_test" checks that the database is mapping to the correct foreign 
key field as defined in the dbdictionary. The "on_update_test" and 
"on_delete_test" make sure the on_update and on_delete fields match between 
the database schema and the dictionary. 
The column "errors" will identify if any of the tests fail for each table 
and field that contains a foreign key.

See the "database" learnr tutorial for more details on the
database. The tutorial can be launched with
\code{learnr::run_tutorial(name = "database", package = "AMMonitor")}.
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in a temporary directory
# (to be deleted)

# run the function and capture the connection
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# list the tables in the AMMonitor database
DBI::dbListTables(conx)

# check the database; do not disconnect on exit
results <- dbCheckup(con = conx, errors_only = TRUE, disconnect = FALSE)
 
# the errors, if present, will be given in the captured output
str(results)

# -----------------------------------------------------
# the demo has no errors
# let's change some data entries so you can see an error returned

# note that we cannot add foreign key errors into the database using the 
# DBI package due to foreign key constraints.


# produce column mismatch
DBI::dbExecute(
 conx, 
 statement = "UPDATE dbdictionary SET pk_fieldname = 'primary' 
   WHERE pk_fieldname = 'primary_account';"
 ) 

# check the database again; disconnect from database
results <- dbCheckup(con = conx, errors_only = TRUE, disconnect = TRUE)

# look at the structure of the results
str(results)
# -----------------------------------------------------

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)


}
}
\seealso{
Other database: 
\code{\link{addRecord}()},
\code{\link{dbAddUserCol}()},
\code{\link{dbAddUserTable}()},
\code{\link{dbCheckCore}()},
\code{\link{dbCheckData}()},
\code{\link{dbCreate}()},
\code{\link{dbImportCSV}()},
\code{\link{dbUpdateVersion}()},
\code{\link{deleteRecord}()}
}
\concept{database}
