% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreate.R
\name{dbCreate}
\alias{dbCreate}
\title{Create a new AMMonitor SQLite database}
\usage{
dbCreate(new_db_name, new_db_filepath, db_source = "default", 
  source_filepath = NA)
}
\arguments{
\item{new_db_name}{Name of the database.}

\item{new_db_filepath}{File path to the directory that will house the
database.  This should be the "database" directory of an AMMonitor project.}

\item{db_source}{Source of the database dictionary and lists (options 
include "default", "existing_db", "sciencebase"). The "default" option
should be used to create a new database from scratch.}

\item{source_filepath}{If db_source = "existing_db" or "sciencebase", 
provide
the file path to the existing database or science base volume.}
}
\value{
A new AMMonitor SQLite database.
}
\description{
Create a new AMMonitor SQLite database.  All tables and keys 
are generated automatically. The "default" database is a new database
populated with default values and ready to intake new information. See 
details.
}
\details{
The database should reside in the AMMonitor project (the working 
directory) within the directory named "database".

This function should be called directly only in cases where a new,
default AMMonitor database is to be created. The default dictionary comes 
with the AMMonitor package, where the
dictionary provides all information needed to create a new SQLite database.
If creating a new AMMonitor database from a data release published in the 
USGS ScienceBase repository, please use the \code{sbRehydrate()} function
instead.

Users can interact with their database in two ways: 1) With R code;
2) with the AMMonitor RShiny front end. Launch the Shiny application with 
the \code{launchApp()} function.

See the "database" learnr tutorial for more details on the database.
The tutorial can be launched with \code{learnr::run_tutorial(name = "database",
package = "AMMonitor")}.
}
\examples{
\dontrun{

# ------------------------------------------------------------
# Create a new default AMMonitor database for demonstration
# in your temporary directory (to be deleted).
# ------------------------------------------------------------

# create the default database (to be deleted):
new_db_filepath <- tempdir()
new_db_name <- "demo.sqlite"
db_source <- "default"
source_filepath <- NULL

# create the new SQLite database
dbCreate(new_db_name, new_db_filepath, db_source, source_filepath)

# verify that the database exists in your current working directory
file.exists(new_db_name)

# to work with a database in code, you must set a connection
conx <- dbSetCon(paste0(new_db_filepath, "/", new_db_name))

# look at the connection
conx

# look at the built-in lists table
DBI::dbGetQuery(conn = conx, statement = "SELECT * FROM lists")

# look at the built-in people table; notice the default unknownPerson
DBI::dbGetQuery(conn = conx, statement = "SELECT * FROM people")

# add a record to person table
rs <- DBI::dbExecute(
 conn = conx,
 statement = "
  INSERT INTO people(pk_personid, first_name, last_name,
    project_role, email, phone, sb_login, display_name)
  VALUES('fbaggins', 'Frodo', 'Baggins', 'Technician', 'frodo@gmail.com',
   '800-THE-RING', 'NA', 'Frodo');"
)

# look at the updated people table
DBI::dbGetQuery(conn = conx, statement = "SELECT * FROM people")

# look at the dbdictionary table; it provides information about the 
# database schema
View(DBI::dbGetQuery(conn = conx, statement = "SELECT * FROM dbdictionary"))

# disconnect from database when finished
DBI::dbDisconnect(conx)

# delete the test database and remove it from your working directory
unlink(new_db_name)

}

}
\seealso{
Other database: 
\code{\link{addRecord}()},
\code{\link{dbAddUserCol}()},
\code{\link{dbAddUserTable}()},
\code{\link{dbCheckCore}()},
\code{\link{dbCheckData}()},
\code{\link{dbCheckup}()},
\code{\link{dbImportCSV}()},
\code{\link{dbUpdateVersion}()},
\code{\link{deleteRecord}()}
}
\concept{database}
