% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDistributed.R
\name{dbDistributed}
\alias{dbDistributed}
\title{Create a stand-alone database for distribution to
volunteer taggers or verifiers}
\usage{
dbDistributed(con, out_filepath, person_id, media_ids, mode = "verify",
use_renv = FALSE, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{out_filepath}{The path to store the mini project}

\item{person_id}{The tagger/verifier's pk_personid from the people table.}

\item{media_ids}{Vector of pk_mediaids from the media table.}

\item{mode}{"annotate" or "verify".}

\item{use_renv}{TRUE or FALSE. Should renv be used to create the distributed 
file? Currently unsupported and default is FALSE.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE.}
}
\value{
A directory of a mini AMMonitor project that can be compressed and 
distributed to taggers/verifiers for independent work
}
\description{
Function to be used by AMMonitor database managers to create a
stand-alone mini AMMonitor project in a specified directory for
distribution to others to permit external collaborators to tag or verify
media without access to the master AMMonitor database.
}
\details{
See the "dbdistributed" learnr tutorial for more details on the use
of this function.

The tutorial can be launched with learnr::run_tutorial(name = "dbdistributed",
package = "AMMonitor").
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the annotations table; Bilbo made annotations; Frodo will verify
# them
DBI::dbReadTable(conx, name = "annotations")

# create a mini AMMonitor project to send to Frodo
distributed_fp <- dbDistributed(
 con = conx,
 out_filepath = tempdir(),
 person_id = "fbaggins", 
 media_ids = 1:15, 
 mode = "verify", 
 use_renv = FALSE,
 disconnect = FALSE)
 
# go look at the newly created mini-project; zip and send to Frodo
 dir.exists(distributed_fp)
 
# show the files in the mini project
 list.files(distributed_fp, recursive = TRUE)
 
# launch the mini project in Shiny 
 launchApp(distributed_fp)
 
# another example; let Gandalf make annotations too
 distributed_fp <- dbDistributed(
 con = conx,
 out_filepath = tempdir(),
 person_id = "gandalf", 
 media_ids = 1:15, 
 mode = "annotate", 
 use_renv = FALSE,
 disconnect = FALSE)
 
 # go look at the newly created mini-project
 dir.exists(distributed_fp)
 
 # test the mini project by launching the mini project in Shiny 
 launchApp(distributed_fp)
 
 # if all ok, zip and send to Gandalf
}

}
\seealso{
Other annotation: 
\code{\link{dbDistributedImport}()}
}
\concept{annotation}
