% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDistributedImport.R
\name{dbDistributedImport}
\alias{dbDistributedImport}
\title{Import data from a distributed database}
\usage{
dbDistributedImport(con, con_distributed, mode, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to the master database}

\item{con_distributed}{An open connection to the distributed database}

\item{mode}{What data should be imported: "annotate" or "verify"}

\item{disconnect}{TRUE or FALSE. Should the master database connection be severed
on exit? Default is FALSE}
}
\value{
Message indicating whether new data was successfully imported
}
\description{
Function to be used by AMMonitor database managers to import
externally created annotations or verifications created from the 
\code{dbDistributed} function.
}
\details{
See the "dbdistributed" learnr tutorial for more details on the database.
The tutorial can be launched with learnr::run_tutorial(name = "dbdistributed",
package = "AMMonitor").
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the annotations table; note that Bilbo made 25 annotations
DBI::dbReadTable(conx, name = "annotations")

# look at the verifications table; note that Sam verified several
# of Bilbo's annotations
DBI::dbReadTable(conx, name = "annotationverifications")

# let's let Gandalf verify Bilbo's records too
# create a distributed database for Gandalf to work on; 
# Gandalf will add verifications to media files 1:15

distributed_fp <- dbDistributed(
 con = conx,
 out_filepath = tempdir(),
 person_id = "gandalf", 
 media_ids = 1:15, 
 mode = "verify", 
 use_renv = FALSE,
 disconnect = FALSE)
 
 # make sure the distributed mini project exists
 file.exists(distributed_fp)
 
 # zip and distribute the project to Gandalf to add verifications ----
 # simulate some new verifications by Gandalf
 # we'll assume Gandalf verifies the first 15 annotations as correct annotations
 db_name <- list.files(paste0(distributed_fp, "/database"), full.names = TRUE)
 gandalf_con <- dbSetCon(db_name)
 
 DBI::dbAppendTable(
  conn = gandalf_con,
  name = "annotationverifications",
  value = data.frame(
    pk_annoverificationid = NA,
    is_valid = 1,
    fk_personid = "gandalf",
    fk_annotationid = 1:15,
    timestamp = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S")
  )
 )
 
 # confirm gandalf's records in his database
 DBI::dbReadTable(gandalf_con, name = "annotationverifications")
 
 # import Gandalf's verifications to the master database
 dbDistributedImport(
  con = conx,
  con_distributed = gandalf_con,
  mode = "verify",
  disconnect = FALSE
 )
 
 # check the master annotationverifications table
 DBI::dbReadTable(conx, "annotationverifications")

}

}
\seealso{
Other annotation: 
\code{\link{dbDistributed}()}
}
\concept{annotation}
