% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetSummaryData.R
\name{dbGetSummaryData}
\alias{dbGetSummaryData}
\title{Creates a list of dataframes from a database for summary purposes}
\usage{
dbGetSummaryData(con, m_dates = c(NA, NA), disconnect = FALSE)
}
\arguments{
\item{con}{An open database connection}

\item{m_dates}{A vector of the start and end dates of media to be 
included in the summary.
Default is NA, which summarizes the entire database.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE}
}
\description{
Creates a list of dataframes suitable for passing into the 
\code{dbTableSummary()} or \code{dbPlotSummary()} functions.
}
\details{
\code{dbGetSummaryData()} generates a list of database dataframes that
is suitable for passing to the \code{dbPlotSummary()} or \code{dbTableSummary()}
functions. The returned object may be quite large, but enables users to work 
with database information with R functions rather than making SQL calls to the 
database. 

The dataframes returned include:
\itemize{
  \item \strong{media}: Dataframe that contains information on the media 
  collected within the specified time span.
  \item \strong{field_visits}: Dataframe that contains information
  on field visits done within the specified time span.
  \item \strong{media_visits}: Dataframe that contains information 
  that links media dates with visit dates.
  \item \strong{locations}: Dataframe that contains information on
  locations associated with media files collected within the 
  specified time span.
  \item \strong{annotations}: Dataframe containing information on
  annotations made on media files collected within the specified
  time span.
  \item \strong{taxa}: Dataframe containing information on taxa
  associated with annotations made on media files collected within the 
  specified time span.
  \item \strong{modeloutputs}: Dataframe containing information on
  on machine learning model outputs associated with media files collected
  within the specified time span.
}

See the "dbsummary" learnr tutorial for more details on the
dbGetSummaryData() function. The tutorial can be launched with
\code{learnr::run_tutorial(name = "dbsummary", package = "AMMonitor")}.
}
\examples{
\dontrun{

# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# look at the folders within an AMMonitor project; notice the database folder
list.files(demo_fp, recursive = FALSE)

# the sample database is named "demo.sqlite"
list.files(file.path(demo_fp, "database"))

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# get summary data and disconnect from database
results <- dbGetSummaryData(con = conx, disconnect = FALSE)
 
# look at the returned object; it is a list of dataframes
str(results, max.level = 1)

# look at the first 3 records of each dataframe
lapply(results,  FUN = head, n = 3)

# an example with date subsetting ----------------------

# get summary data within specified data range and disconnect from database
results <- dbGetSummaryData(
  con = conx,
  m_dates = c("2023-05-01", "2023-12-31"),
  disconnect = TRUE
)

str(results, max.level = 1)

# -----------------------------------------------------

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)

} 


}
\concept{annotations}
\concept{locations}
\concept{media}
\concept{modeloutputs}
\concept{people}
\concept{taxa}
\concept{visits}
