% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbImportCSV.R
\name{dbImportCSV}
\alias{dbImportCSV}
\title{Import records from a csv into an AMMonitor table}
\usage{
dbImportCSV(con, csv_path, table_name, row_numbers, disconnect = FALSE)
}
\arguments{
\item{con}{An open database connection}

\item{csv_path}{File path to a csv containing rows to be added}

\item{table_name}{The name of the table to import records into}

\item{row_numbers}{A vector of row numbers to import, if only a subset is
to be imported. Default is NA, which will import all of the records in the 
CSV file}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed?
Default is FALSE.}
}
\value{
A data.frame that stores the import status for each record in the CSV
}
\description{
Import records from a csv into an AMMonitor table
}
\examples{
\dontrun{

# create a new database in the temporary directory
dbCreate(new_db_name = "demo.sqlite", new_db_filepath = tempdir())

# run dbSetCon to set and store the open connection
conx <- dbSetCon(dbPath = file.path(tempdir(), "demo.sqlite"))

# write a csv with a record to add to the "people" table
csv_path <- file.path(tempdir(), 'test.csv')

utils::write.csv(
  x = data.frame(
    pk_personid = c("bbaggins", "fbaggins"),
    first_name = c("Bilbo", "Frodo"),
    last_name = c("Baggins", "Baggins"),
    project_role = c("Tagger", "Field technician"),
    display_name = c("blue", "red")), 
  file = csv_path
)

# import records from the CSV file into the people table
rs <- dbImportCSV(
  con = conx,
  csv_path = csv_path,
  table_name = "people",
  disconnect = FALSE
)

# inspect the returned results
rs

# check to see that the new records were added:
DBI::dbReadTable(conx, "people")

# disconnect from the database and clean up temporary files
DBI::dbDisconnect(conx)

unlink(file.path(tempdir(), "demo.sqlite"))
unlink(csv_path)

}

}
\seealso{
Other database: 
\code{\link{addRecord}()},
\code{\link{dbAddUserCol}()},
\code{\link{dbAddUserTable}()},
\code{\link{dbCheckCore}()},
\code{\link{dbCheckData}()},
\code{\link{dbCheckup}()},
\code{\link{dbCreate}()},
\code{\link{dbUpdateVersion}()},
\code{\link{deleteRecord}()}
}
\concept{database}
