% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbPlotSummary.R
\name{dbPlotSummary}
\alias{dbPlotSummary}
\title{Creates database summary figures}
\usage{
dbPlotSummary(
 summary_data,
 num_bins = 15,
 map_database = NA,
 regions = NA,
 key_species = NULL,
 show_plots = FALSE)
}
\arguments{
\item{summary_data}{A list of dataframes generated from the 
\code{dbGetSummaryData()} function}

\item{num_bins}{Number of bins to break the date columns into; default is 
15 for  plotting.  Also could  enter "month", "year", or "quarter".}

\item{map_database}{Name of map database for plotting locations. Options 
are
 "world" or "usa" or "state" or "county"}

\item{regions}{Vector of regions within the map_database for plotting 
(e.g., c("vermont"))}

\item{key_species}{Vector of taxa ids to include for additional plots and
maps}

\item{show_plots}{TRUE or FALSE. Should the plots be printed? 
Default is FALSE}
}
\description{
Inputs result from \code{dbGetSummaryData()} and creates 
database 
summary figures for inspection of outliers and for use in reporting
}
\details{
\code{dbPlotSummary()} provides graphical summaries of different 
elements 
of an AMMonitor database.  A primary input is a list generated from the 
function \code{dbGetSummaryData()}, which extracts portions of the database 
for
inspection and summary.   The primary output is a list of ggplot objects
created by the package, ggplot2.  Each output can be further modified by 
the user with ggplot2 code.

Map boundaries are produced from the package, maps.  The 'map_database' 
argument in \code{dbPlotSummary()} 
feeds the 'database' argument to the \code{maps::map()} function.  It is a 
character 
string naming a geographical database. The string choices include a world 
map, 
three USA databases (usa, state, county), and more.  The 'regions'  argument 
in \code{dbPlotSummary()} feeds the 'regions'
 argument from the \code{maps::map()} function.   

In the maps package, each database is 
composed of a collection of polygons, and each polygon has a unique name. 
When a region is composed of more than one polygon, the individual polygons 
have the name of the region, followed by a colon and a qualifier, as in 
michigan:north and michigan:south. Each element of regions is matched 
against 
the polygon names in the database and, according to exact, a subset is 
selected for drawing. The regions may also be defined using (perl) regular 
expressions. This makes it possible to use 'negative' expressions like 
"Norway(?!:Svalbard)", which means Norway and all islands except Svalbard. 
All entries are case insensitive. 
 
 Type \code{help(package = 'maps')} to see the maps package index page. 
 
See the "dbsummary" learnr tutorial for more details on the
dbPlotSummary() function. The tutorial can be launched with
\code{learnr::run_tutorial(name = "dbsummary", package = "AMMonitor")}.
}
\examples{
\dontrun{

# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# the sample database is named "demo.sqlite" within the database directory
list.files(file.path(demo_fp, "database"))

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# get summary data and disconnect from database
results <- dbGetSummaryData(con = conx, disconnect = TRUE)
 
# look at the returned object; it is a list of dataframes
str(results, max.level = 1)

# load ggplot2
library(ggplot2)

# send the captured outputs to dbPlotSummary to generate summary plots
plot_results <- dbPlotSummary(
  summary_data = results,
  num_bins = 'month',
  map_database = "state", 
  regions = "vermont",
  key_species = c("moose", "black bear"),
  show_plots = FALSE
)
 
# the returned object is a list of ggplot objects
str(plot_results, max.level = 1)

# or just look at the names of the ggplot objects
names(plot_results)

#  see all results in plot window
plot_results

# or plot individually for inspection
plot_results[[1]]

# each element can be retrieved by name or index and edited
plot1 <- plot_results[[1]]
plot(plot1)

# customize plot1 to your liking
plot1 <- plot1 +
 scale_fill_viridis_d()
plot1
 
# make sure to update the list if you want to save plots
plot_results[['gg_media_month_year_hist']] <- plot1

# confirm the update
plot_results[[1]]

} 


}
