% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSetCon.R
\name{dbSetCon}
\alias{dbSetCon}
\title{Set a SQLite connection to a database and turn on foreign keys}
\usage{
dbSetCon(dbPath)
}
\arguments{
\item{dbPath}{Path to SQLite database}
}
\value{
An object that stores the open database connection
}
\description{
Creates a SQLite connection to a database and turns on foreign keys
}
\examples{
\dontrun{

# create an new database in the working directory
dbCreate(new_db_name = "demo.sqlite", new_db_filepath = tempdir())

# run dbSetCon to set and store the open connection
con <- dbSetCon(dbPath = file.path(tempdir(), "demo.sqlite"))

# look at the stored connection object
class(con)
con

}
}
