% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableSummary.R
\name{dbTableSummary}
\alias{dbTableSummary}
\title{Creates database summary figures}
\usage{
dbTableSummary(summary_data, key_species = NULL)
}
\arguments{
\item{summary_data}{A list of dataframes generated from dbGetSummaryData()}

\item{key_species}{A vector of up to ten key taxa to include for annotations 
and verifications tables. Default includes all taxa.}
}
\description{
Creates database summary tables
for inspection of outliers and for use in reporting
}
\details{
dbTableSummary() provides table summaries of different elements 
of an AMMonitor database.  A primary input is a list generated from the 
function \code{dbGetsummaryData()}, which extracts portions of the database for
inspection and summary. 
The output is a list of gt table objects, which can be  
manipulated by the user using the gt package to create professional tables.

See the "dbsummary" learnr tutorial for more details on the
dbTableSummary() function.  The tutorial can be launched with
\code{learnr::run_tutorial(name = "dbsummary", package = "AMMonitor")}.
}
\examples{
\dontrun{

# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# the sample database is named "demo.sqlite" within the database directory
list.files(file.path(demo_fp, "database"))

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# get summary data and disconnect from database
results <- dbGetSummaryData(con = conx, disconnect = FALSE)
 
# look at the returned object; it is a list of dataframes
str(results, max.level = 1)

# generate summary tables
table_results <- dbTableSummary(
 summary_data = results,
 key_species = c("moose", "black bear")
 )
 
str(table_results, max.level = 1)

# each output is saved in a list element that can be retrieved and edited

# let's take a look at the collected media files by year and location
collected_media <- table_results$media_collected_per_year_location

# look at this table
collected_media

# change column names
collected_media <- collected_media |> 
  gt::cols_label(
    year = "Year",
    fk_locationid = "Location ID",
    collected_media_files = "Collected Media Files") 
    
# add source note
collected_media <- collected_media |>
  gt::tab_source_note(
    source_note = "Source: AMMonitor Demo Dataset")
    
# look at the updated table
collected_media

# save the table back to the list if desired
table_results$media_collected_per_year_location <- collected_media

# refer to the gt package for more ideas on manipulating your results tables 



} 


}
