% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUpdateVersion.R
\name{dbUpdateVersion}
\alias{dbUpdateVersion}
\title{Update AMMonitor databases to the latest version}
\usage{
dbUpdateVersion(db_path, from_version, to_version, dest = NA)
}
\arguments{
\item{db_path}{A file path to an AMMonitor SQLite database}

\item{from_version}{Version of the existing AMMonitor database to convert from}

\item{to_version}{New AMMonitor database version to convert to}

\item{dest}{Directory path for where to save the updated database. Defaults
to the directory containing the unconverted database \code{(dirname(db_path))}}
}
\value{
File path to the newly created, updated database.
}
\description{
Updates an AMMonitor SQLITE database to match the schema of a 
given AMMonitor package version.
}
\details{
This function can be used to update an AMMonitor SQLite database. In
some new versions, the database schema is slightly altered, and this function
allows a user to specify the current version of their AMMonitor database and
the desired, updated version. The original database will not be altered and
a new database, updated database will be created. It is recommended that 
users use the latest released version of AMMonitor and update their database 
accordingly.

One important caveats is that before using this function, it is highly
recommended that you confirm database integrity before proceeding. You can 
check the status of your database with the AMMonitor functions `dbCheckCore()`,
`dbCheckData()`, and `dbCheckup()`. 

This function was introduced in AMMonitor version 2.2 and currently
only has the ability to update a database from version 2.1 to 2.2. If you have
used either the *monitoR* database or the database from the first release of 
*AMMonitor*, please contact one of the package authors for advice on how to 
convert your database to the latest version. 

See the "database" learnr tutorial for more details on the
database. The tutorial can be launched with
\code{learnr::run_tutorial(name = "database", package = "AMMonitor")}.
}
\examples{
\dontrun{

Download the version 2.1 demo database (to be deleted)
db_path_original <- file.path(tempdir(), "demo_v21.sqlite")

utils::download.file(
  url = "https://code.usgs.gov/vtcfwru/ammonitor/-/raw/AMMonitor2.1/inst/extdata/demoAMM/database/demo.sqlite", 
  destfile = db_path_original, 
  mode = "wb"
)

# Connect to the demo database and explore some V2.1 distinguishing features
conx <- AMMonitor::dbSetCon(db_path_original)

# Should be FALSE, since modellables table introduced in V2.2
"modellabels" \%in\% DBI::dbListTables(conx)

DBI::dbDisconnect(conx)

# Update the database from V2.1 to V2.2
db_path_updated <- AMMonitor::dbUpdateVersion(
  db_path = db_path_original,
  from_version = "2.1",
  to_version = "2.2",
  dest = tempdir()
)

# Check that the updated db contains V2.2 features
conx <- AMMonitor::dbSetCon(db_path_updated)

# Should be TRUE, since modellables table introduced in V2.2
"modellabels" \%in\% DBI::dbListTables(conx)

DBI::dbDisconnect(conx)

# Clean-up
unlink(db_path_original)
unlink(db_path_updated)

}
}
\seealso{
Other database: 
\code{\link{addRecord}()},
\code{\link{dbAddUserCol}()},
\code{\link{dbAddUserTable}()},
\code{\link{dbCheckCore}()},
\code{\link{dbCheckData}()},
\code{\link{dbCheckup}()},
\code{\link{dbCreate}()},
\code{\link{dbImportCSV}()},
\code{\link{deleteRecord}()}
}
\concept{database}
