% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteRecord.R
\name{deleteRecord}
\alias{deleteRecord}
\title{Deletes a record from a specified table and/or any associated files}
\usage{
deleteRecord(con, table_name, selected_row, delete_file = FALSE, 
  media_paths = NA, amm_project_path = getwd())
}
\arguments{
\item{con}{A connection to the database}

\item{table_name}{A string, the name of the table to delete a record from}

\item{selected_row}{The contents of the selected row (dataframe)
or character vector (vector of primary key values, in the order that they
are in the table)}

\item{delete_file}{TRUE or FALSE, specifying whether to delete associated files, 
if any. Default is FALSE}

\item{media_paths}{Media paths (for use in AMMonitor's Shiny app, default = NA)}

\item{amm_project_path}{Filepath to the outermost directory for an AMMonitor
project}
}
\value{
Returns a list with the status of the deletion (TRUE if successful) and
a message indicating any errors encountered, if the deletion failed.
}
\description{
Deletes a record from a table and its associated file
(if applicable) in local, Google Drive, AWS S3, or SharePoint directories.
}
\details{
When deleting a record from the AMMonitor database, the media, models, 
logs, and spatials tables all reference external files, which are tracked with 
the filepath field (media, spatials), amml field (models), or pk_logid field 
(logs). This delete function performs a regular delete query for all other tables, 
but for these four, this function attempts to delete the associated file before 
deleting the database record, when specified, based on the associated file path.
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in a temporary directory
# (to be deleted)

# run the function and capture the connection
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# add and person to the database
new_person <- data.frame(
  pk_personid = 'test_person', 
  display_name = 'test')
  
DBI::dbAppendTable(conx, 'people', new_person)

# look at the new addition
DBI::dbReadTable(conx, name = "people")

# now delete that record, referencing the table and primary key
rs <- deleteRecord(
  con = conx, 
  table_name = 'people', 
  selected_row = new_person)

# print the status of the deletion
print(rs)

# delete a media record from the demo db, along with the associated file
# get the media file to be deleted
media_to_delete <- DBI::dbGetQuery(con, 'SELECT * FROM media WHERE pk_mediaid = 1;')

# check to see that the associated file exists
(file.exists(
  paste(demo_fp, paste0(
    media_to_delete$media_type, 's'), media_to_delete$filename, sep = '/')
  )
 )

rs <- deleteRecord(
  con = conx, 
  table_name = 'media', 
  selected_row = media_to_delete , 
  delete_file = TRUE, 
  amm_project_path = demo_fp
)

# print the status of the deletion
print(rs)

# check to see that the associated file was deleted
file.exists(
  paste(
    demo_fp, 
    paste0(media_to_delete$media_type, 's'), media_to_delete$filename, sep = '/')
 )

# disconnect from the database when finished
DBI::dbDisconnect(con)

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)

}
}
\seealso{
Other database: 
\code{\link{addRecord}()},
\code{\link{dbAddUserCol}()},
\code{\link{dbAddUserTable}()},
\code{\link{dbCheckCore}()},
\code{\link{dbCheckData}()},
\code{\link{dbCheckup}()},
\code{\link{dbCreate}()},
\code{\link{dbImportCSV}()},
\code{\link{dbUpdateVersion}()}
}
\concept{database}
