% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equipAddConfig.R
\name{equipAddConfig}
\alias{equipAddConfig}
\title{Adds an equipment configuration to an AMMonitor database}
\usage{
equipAddConfig(con, excel_fp, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{excel_fp}{A filepath to an Excel file containing new equipment
configuration information.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
An updated database containing the new equipment configuration
}
\description{
This function adds a new equipment configuration to an AMMonitor
database and updates all associated tables.
}
\details{
The equipAddConfig function can be used to easily add new equipment 
configurations to an AMMonitor database through a simple Excel spreadsheet.
This process updates all relevant AMMonitor tables. A spreadsheet can be
generated for a specific equipment model using the equipConfigWriteXL()
function. 

This workflow is useful for adding equipment configurations programmatically,
or for adding a configuration to multiple projects (in this case, check your
keys). Configurations can also be added through the AMMonitor app using the
equipManager app; see the "equipment" tutorial for more information.

If an equipment model isn't in the database yet, see equipModelAdd().

Open your generated spreadsheet and fill out each sheet to create a new
equipment configuration. Each sheet corresponds to a database table as
follows:

Configuration_Name corresponds to the econfignames table. The fk_equipmodelid
should be automatically filled in when the spreadsheet is created. All other
fields should be filled in.

Configuration_Values corresponds to the econfigvalues table. This sheet will
include rows for all settings associated with the equipment model specified,
along with descriptions and Option_Name and Value columns to fill in. If
a setting has options, they will be listed next to the row. Please fill in
the Option_Name and Value columns. Leave an empty cell for blank entries.

Setting_Options provides a list of all valid entries for settings with 
defined options in the database. If a setting has no options listed, it takes
a numeric entry in the Value column. This sheet does not need to be edited, 
but should be used for reference information about any setting options.

For more information about equipment configurations and the associated 
tables, or for a full walkthrough, refer to the "equipment" tutorial, 
which can be run using the following code:
\code{learnr::run_tutorial(name = "equipment", package = "AMMonitor")}
}
\examples{
\dontrun{
# Create a demo database
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Connect to the database
conx <- dbSetCon(file.path(demo_fp, "/database/demo.sqlite"))

# Take a look at the equipmodels table
DBI::dbReadTable(conx, "equipmodels")

# Let's make a spreadsheet for recorder_model_x and write it to the demo dir
equipConfigWriteXL(
  con = conx,
  equipmodelid = "recorder_model_x",
  output_dir = demo_fp,
  disconnect = FALSE)

# Fill in the spreadsheet with equipment configuration info
print(paste(
  "To complete this demo, open the excel file at",
  file.path(demo_fp, "equipAddConfig.xlsx"),
  "and add the configuration info.",
  "Note: the model name and config options are",
  "pre-populated in this spreadsheet."
))

# Add the new configuration to the database
equipAddConfig(
  con = conx,
  excel_fp = file.path(demo_fp, "equipAddConfig.xlsx"),
  disconnect = FALSE)
  
# Retrieve config name from spreadsheet
config_name <- readxl::read_excel(
  file.path(demo_fp, "equipAddConfig.xlsx"), 
  sheet = "Configuration_Name", 
  col_names = FALSE, 
  range = "A2"
)
  
# Query the database for your new configuration
qryEconfiguration(
  con = conx,
  econfigname = config_name,
  disconnect = FALSE)
  
# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)
}

}
