% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equipConfigWriteXL.R
\name{equipConfigWriteXL}
\alias{equipConfigWriteXL}
\title{Writes an Excel spreadsheet to be used to easily add a new equipment
model configuration to an AMMonitor database}
\usage{
equipConfigWriteXL(con, equipmodelid, output_dir, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{equipmodelid}{The pk_equipmodelid of the equipment model you'd like to add a
configuration for, as listed in the "equipmodels" table}

\item{output_dir}{A filepath to the directory where you'd like the output
Excel spreadsheet to be saved.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
An Excel spreadsheet containing tables to add a new equipment model 
configuration to the database in the specified directory
}
\description{
This function writes an Excel spreadsheet to a provided 
directory which can be used to easily add a new equipment configuration to an
AMMonitor database. This spreadsheet contains multiple sheets to specify
configuration names and values, as well as listing all valid settings for the
provided equipment model. For more information, see the Details section.
}
\details{
This function creates an Excel spreadsheet to enable the easy addition of new
equipment configuration to an AMMonitor database. It queries the database
based on the provided equipment model ID to output a spreadsheet that 
includes relevant settings and options for the selected model. This
spreadsheet can be completed and input into equipAddConfig() to update the
database with a new equipment configuration.

New equipment configurations can also be added using the equipManager app in
AMMonitor. For more information on using the app, see the "equipment" 
tutorial.

The output spreadsheet will be named "equipAddConfig.xlsx" and stored in the
specified output directory. It contains three sheets:

Configuration_Name corresponds to the econfignames table. The fk_equipmodelid
will be automatically filled in based on the equipment model specified in
this function. All other details can be filled in for the new configuration.

Configuration_Values corresponds to the econfigvalues table. This sheet will
include rows for all settings associated with the equipment model specified,
along with descriptions and Option_Name and Value columns to fill in.

Setting_Options provides a list of all valid entries for settings with 
defined options in the database. If a setting has no options listed, it takes
a numeric entry in the Value column.

For information on adding a completed spreadsheet to the database, see
equipAddConfig(). For a full walkthrough, see the "equipment" tutorial,
which can be launched with the following code:
\code{learnr::run_tutorial(name = "equipment", package = "AMMonitor")}
}
\examples{
\dontrun{
# Create a demo database
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Connect to the database
conx <- dbSetCon(file.path(demo_fp, "/database/demo.sqlite"))

# Take a look at the equipmodels table
DBI::dbReadTable(conx, "equipmodels")

# Let's make a spreadsheet for recorder_model_x and write it to the demo dir
equipConfigWriteXL(
  con = conx,
  equipmodelid = "recorder_model_x",
  output_dir = demo_fp,
  disconnect = FALSE)
}

}
