% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equipModelAdd.R
\name{equipModelAdd}
\alias{equipModelAdd}
\title{Adds an equipment model, settings, and options to an AMMonitor
database}
\usage{
equipModelAdd(con, excel_fp, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{excel_fp}{A filepath to an Excel file containing new equipment model 
information.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
An updated database with new equipment model info
}
\description{
This function adds a new equipment model, along with any 
settings and associated options, to an AMMonitor database using an Excel
spreadsheet. See Details for full instructions.
}
\details{
The equipModelAdd() function simplifies the process of adding new equipment
models, settings, and options to an AMMonitor database using an Excel
spreadsheet. This approach can be useful for adding equipment models
programmatically, or for adding an equipment model to multiple projects. New
equipment models can also be added using the equipManager app; for more
information, see the "equipment" tutorial.

A template spreadsheet is provided and can be copied from the
package to a directory of your choice using the following code:

\code{fp <- system.file("extdata/equipModelAdd.xlsx", package = "AMMonitor")}
 
\code{file.copy(from = fp, to = tempdir(), overwrite = FALSE)}

Open the copied Excel file and fill out the listed fields for the equipment
model you'd like to add. Do not alter the original spreadsheet; AMMonitor
expects a specific format to successfully add the new entries to the 
database. Each table in the spreadsheet corresponds to a database table:

Table 1: EquipModel (Blue) corresponds to the equipmodels table.

Table 2: Settings (Green) corresponds to the esettingnames table.

Table 3: Options (Orange) corresponds to the esettingoptions table.

Fields in the spreadsheet represent the database columns of the same names.
For more information about these tables and adding equipment models to an
AMMonitor database, see the "equipment" tutorial. It can be accessed with the
following code:
\code{learnr::run_tutorial(name = "equipment", package = "AMMonitor")}
}
\examples{
\dontrun{

# Create a demo database
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Connect to the database
conx <- dbSetCon(file.path(demo_fp, "/database/demo.sqlite"))

# Copy the template spreadsheet to your demo directory
fp <- system.file("extdata/equipModelAdd.xlsx", package = "AMMonitor")
file.copy(from = fp, to = demo_fp, overwrite = FALSE)

# Fill in the tables with example equipment model information

# Add the model to the database
equipModelAdd(
  con = conx,
  excel_fp = file.path(demo_fp, "equipModelAdd.xlsx"),
  disconnect = FALSE)
  
# Check the updated equipmodels table
DBI::dbReadTable(conx, "equipmodels")

# Check the updated esettingnames table
DBI::dbReadTable(conx, "esettingnames")

# Check the updated esettingoptions table
DBI::dbReadTable(conx, "esettingoptions")

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)
}

}
