% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoresDetect.R
\name{formatDetections}
\alias{formatDetections}
\title{Formats detections from a detectionList object to fit the AMMonitor database}
\usage{
formatDetections(template_info, recording_info, detections)
}
\arguments{
\item{template_info}{A dataframe containing information about templates stored in the AMMonitor database}

\item{recording_info}{A one-row dataframe containing information about an audio recording stored in the AMMonitor database.}

\item{detections}{A detectionList object, returned after running binMatch or corMatch}
}
\description{
An internal function formats detectionList detections for insertion into the AMMonitor database.
}
