% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoresDetect.R
\name{getAudioIDs}
\alias{getAudioIDs}
\title{Get media IDs and file names for specific audio files}
\usage{
getAudioIDs(con, recordingNames = "all")
}
\arguments{
\item{con}{A connection to a SQLite or Postgres AMMonitor database.}

\item{recordingNames}{Default = 'all'. An optional character vector specifying
by name which recordings should be processed. Default is 'all'.}
}
\description{
An internal function that retrieves media IDs and audio files from an AMMonitor database.
}
