% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierPredict.R
\name{getClassifierInfo}
\alias{getClassifierInfo}
\title{Get information about a classifier}
\usage{
getClassifierInfo(con, classifierName)
}
\arguments{
\item{con}{A connection to a SQLite or Postgres AMMonitor database.}

\item{classifierName}{A character vector specifying which classifier to run on template detections.}
}
\description{
An internal function that retrieves information about a classifier from an AMMonitor database.
}
