% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierPredict.R
\name{getTemplateDetections}
\alias{getTemplateDetections}
\title{Retrieves detections of a particular template from an AMMonitor database}
\usage{
getTemplateDetections(
  con,
  templateID,
  recordingRootPath,
  recordingIDs = NULL,
  showProgress = FALSE
)
}
\arguments{
\item{con}{A connection to an AMMonitor database.}

\item{templateID}{The integer pk_templateID of a template in an AMMonitor database.}

\item{recordingRootPath}{A character string indicating the path to the directory containing the recordings, can be local or a URL.}

\item{recordingIDs}{An integer vector of recording IDs to filter detections by.
default is NULL, which gets all detections.}

\item{showProgress}{Default = FALSE. If TRUE, progress will be reported during the score calculations.}
}
\description{
An internal function that retrieves a particular template's detections from an AMMonitor database.
}
