% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoresDetect.R
\name{getTemplateIDs}
\alias{getTemplateIDs}
\title{Get media IDs and file names for specific audio files}
\usage{
getTemplateIDs(con, templateNames = "all")
}
\arguments{
\item{con}{A connection to a SQLite or Postgres AMMonitor database.}

\item{templateNames}{Default = 'all'. An optional character vector specifying which templates to use for searching recordings.}
}
\description{
An internal function that retrieves media IDs and audio files from an AMMonitor database.
}
