% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAppCreate.R
\name{googleAppCreate}
\alias{googleAppCreate}
\title{Creates a data source spreadsheet for Google AppSheet from 
an AMMonitor database}
\usage{
googleAppCreate(con, equip_type = 'all', disconnect = FALSE)
}
\arguments{
\item{con}{Open connection to an AMMonitor database}

\item{equip_type}{Type of equipment to include from 
the database. Options are 'camera', 'recorder', or 'all' ('all' is 
the default).}

\item{disconnect}{TRUE or FALSE. Should the 
database connection be severed on exit? 
Default is FALSE}
}
\value{
filepath to the data source .xlsx file
}
\description{
This function creates an Excel workbook from a 
given AMMonitor database that can be used to generate a 
new visit survey in Google AppSheet. The spreadsheet is created in
your working directory.

This function should be used after updating the database 
with any new locations, people, equipment, user fields, etc. 
This will prevent data entry errors 
where a new location, person, etc. is not available for 
survey users. 
 
See details for instructions on creating the survey app 
in Google AppSheet
}
\details{
This function writes an Excel file to the user's working 
directory with all of the information needed to create a cell phone
app in Google AppSheet.
The function will exclude 
foreign key tables and listitems without entries in the database.

 For detailed step-by-step instructions for creating a cell phone app with
 Google AppSheets, please see the
 AMMonitor tutorial for 'mobile_apps' by running
 \code{learnr::run_tutorial(name = "mobile_apps", package = "AMMonitor")}
}
\examples{
\dontrun{

# Create a mini demo directory and database
demo_fp <- ammCreateMiniDemo(tempdir())

# set path to the database
dbPath <- paste0(demo_fp, "/database/demo.sqlite")

# create a connection to the database
conx <- dbSetCon(dbPath)

# create the spreadsheet and capture the filepath
filepath <- googleAppCreate(
  con = conx,
  equip_type = 'camera'
)

# the file is written to your working directory in R
filepath

# now read the mobile_apps tutorial!
}
}
