% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\alias{launch_app}
\title{Launch the Shiny app for AMMonitor}
\usage{
launchApp(amm_project_path = getwd(), ...)
}
\arguments{
\item{amm_project_path}{Filepath to the outermost directory for an AMMonitor
project}

\item{...}{Additional arguments to be passed to the app.}
}
\description{
Launches the main Shiny app for AMMonitor
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in a temporary directory
# (to be deleted)

# run the function and capture the connection
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# launch the shiny app
launchApp(amm_project_path = demo_fp)

# the database connection is automatically severed upon exit


}

}
\keyword{misc}
