% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierPredict.R
\name{loadModels}
\alias{loadModels}
\title{Get models from one or more AMModels libraries}
\usage{
loadModels(model_info, ammlPath = "/ammls")
}
\arguments{
\item{model_info}{A dataframe containing information about classifier, as stored in the AMMonitor database}

\item{ammlPath}{A character string indicating the path to the directory where AMModels Libraries are stored.}
}
\description{
An internal function that retrieves models from AMModels libraries.
}
