% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoresDetect.R
\name{loadTemplates}
\alias{loadTemplates}
\title{Get templates from one or more AMModels libraries}
\usage{
loadTemplates(template_info, ammlPath = "/ammls")
}
\arguments{
\item{template_info}{A dataframe containing information about templates, as stored in the AMMonitor database}

\item{ammlPath}{A character string indicating the path to the directory where AMModels Libraries are stored.}
}
\description{
An internal function that retrieves templates from AMModels libraries.
}
