% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsGetStations.R
\name{locationsGetStations}
\alias{locationsGetStations}
\title{Retrieves GHCN-D weather station codes based on search criteria}
\usage{
locationsGetStations(amm_fp = getwd(), con, noaa_token = "prompt", startDate = NULL,
   endDate = NULL, bbox = NULL, dbInsert = FALSE, disconnect = FALSE)
}
\arguments{
\item{amm_fp}{Filepath to the AMMonitor project directory.  Default is 
getwd().}

\item{con}{An open connection to the AMMonitor database}

\item{noaa_token}{Token obtained from 
\url{https://www.ncdc.noaa.gov/cdo-web/token}.  Options are "settings" or
"prompt".  Default value is "prompt", which prompts a token entry in the 
R console. If the token is stored
in the "settings" directory of your AMMonitor project as a text file,
enter "settings" for this argument. See details below.}

\item{startDate}{Stations with data after this date will be retrieved,
must be a string in yyyy-mm-dd format}

\item{endDate}{Stations with data before this date will be retrieved, must
be a string in yyyy-mm-dd format}

\item{bbox}{NULL or a vector of numeric values representing a bounding box
to search for weather stations  
in the order east, north, south, west:  minimum longitude, maximum latitude,
minimum latitude,  maximum longitude.
If NULL, a bounding box inferred from the locations table of the database
will be used.}

\item{dbInsert}{TRUE or FALSE, indicating whether or not to write these 
locations to the database. Default is FALSE}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE}
}
\value{
A dataframe containing weather stations contained within the bbox
}
\description{
This function retrieves weather station codes from the NOAA API
based on user entered search criteria.
}
\details{
Weather stations can be searched using either a bounding box or 
based on locations currently in the database.

Obtain a token from https://www.ncdc.noaa.gov/cdo-web/token.
Then, if desired, save this token string in a text file
named 'noaa_token.txt', within the "settings" directory of your AMMonitor
project.  Then, run \code{locationsGetStations()} and set the noaa_token 
argument to "settings"; the function will look for the token there.  If the 
noaa_token argument is "prompt", the function will prompt you
to paste in the token in your console.

This function will retrieve up to 1000 weather station locations at a 
maximum.  If the bbox area of interest is very large (containing more
than 1000 stations), set the bbox to a smaller area and run the function
again.  Some weather stations operate for a short time and are no longer in 
operation. Make sure to set the startDate and endDate to eliminate stations
no longer in operation, if desired.

See the temporals tutorial by running the code:
\code{learnr::run_tutorial(name = "temporals", package = "AMMonitor")}
}
\examples{
\dontrun{

# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the locations table; we are going to add weather stations
DBI::dbReadTable(conx, "locations")
  
# get the weather stations with dbInsert = FALSE - add results to db later
stations <- locationsGetStations(
  amm_fp = demo_fp,
  conx,
  noaa_token = "prompt",
  startDate = NULL,
  endDate = NULL,
  bbox =  c(43.15, -73, 43.5, -72.6),  # minlat, minlong, maxlat, maxlong
  dbInsert = FALSE,
  disconnect = FALSE)
 
# look at returned object
str(stations)
head(stations)
 
# append data if dbInsert was set to false
DBI::dbAppendTable(conx, name = "locations", value = stations)
 
# retrieve the locations table; look for new entries
locations <- DBI::dbReadTable(conx, name = "locations")

# prep data for plotting ---------------
indices <- stats::complete.cases(locations$lat, locations$long, locations$location_type)
locations_xy <- locations[indices, ]

# plot with ggplot2
ggplot2::ggplot(locations_xy,  aes(x = long, y = lat))  +
  geom_point(
    data = locations_xy, 
    aes(x = long, y = lat, shape = location_type, color = location_type), 
    size = 3) +
 coord_fixed() +
 labs(
   title = "Map of Locations",
   x = "Longitude",
   y = "Latitude") +
   theme_minimal()
   
# disconnect from the database when finished
DBI::dbDisconnect(conx)
 
# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)

}

}
\seealso{
Other data: 
\code{\link{temporalsGet}()}
}
\concept{data}
\keyword{data}
