% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationsObscure.R
\name{locationsObscure}
\alias{locationsObscure}
\title{Add a random perturbance to all locations for public display}
\usage{
locationsObscure(con, bbLength = 0.02, overwrite = FALSE, 
disconnect = FALSE)
}
\arguments{
\item{con}{An open database connection}

\item{bbLength}{Length of each edge of the bounding box (in Decimal Degrees).
Default value of 0.02 equates to a bounding box with length ~2.2 km}

\item{overwrite}{TRUE or FALSE. Whether or not to overwrite the previous 
bounding box coordinates. Default is FALSE}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE}
}
\value{
Returns the number of rows from the locations table that have been 
updated
}
\description{
For any location in the locations table with spatial_geometry of
"point", the function will jiggle the point location by a random x,y 
coordinate and will center a bounding box around the jiggled point.
}
\examples{
\dontrun{

# create a demo AMMonitor project in a temporary directory
# (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# look at the folders within an AMMonitor project
list.files(demo_fp, recursive = FALSE)

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the locations table; note the missing bounding boxes (long/lat min/max)
DBI::dbReadTable(conx, "locations")

# add obscured locations for those that require, with default size bounding box
locationsObscure(
  con = conx,
  bbLength = 0.02,
  overwrite = TRUE,
  disconnect = FALSE
)

# look again at the locations table to see the obscured bboxes
locs <- DBI::dbReadTable(conx, "locations")

locs

# remove locations that don't have bboxes (e.g. unknownLocation)
locs <- locs[-which(is.na(locs$lat) | is.na(locs$lat)), ]
 
# plot the location bounding boxes; any messages pertain
ggplot2::ggplot(locs) +
  geom_rect(
     aes(
       xmin = long_min, 
       xmax = long_max, 
       ymin = lat_min, 
       ymax = lat_max), 
     color = "blue", 
     fill = NA) +
  theme_minimal()
  
}  
}
