% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{md_outputs}
\alias{md_outputs}
\title{Sample outputs from MegaDector model run}
\format{
A data frame with 33 rows and 14 variables
}
\usage{
data(md_outputs)
}
\description{
Model outputs from an external analysis of photos by MegaDector
\itemize{
  \item pk_modeloutputid Default is NA (will be given number when imported)
  \item fk_mediaid References pk_mediaid in media table
  \item fk_modelid References pk_modelid in the models table
  \item fk_parentid References a parent modeloutput id from this table.
  \item fk_mconfignameid The model configuration used to obtain the output.
  \item fk_taxonid References a primary key  from the taxa table.
  \item fk_librarylistitemid References a primary key in the librarylistitem
    table.
  \item fk_medialistitemid References a primary key in the medialistitems 
    table.
  \item x_min Bounding box x minimum
  \item x_max Bounding box x maximum
  \item y_min Bounding box y minimum
  \item y_max Bounding box y maximum
  \item value_num Value returned by megadetector (closeness score)
  \item timestamp Timestamp of record entry into the database
}
}
\keyword{datasets}
