% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedAudioCreate.R
\name{mergedAudioCreate}
\alias{mergedAudioCreate}
\title{Create a merged audio file from specified annotations/modeloutputs}
\usage{
mergedAudioCreate(con, annotation_ids = NA, modeloutput_ids = NA,
buffer_before = 5, buffer_after = 5, buffer_btwn = 0.5, sample_rate = 24000,
ammPath = NA, storage_type = "local", media_folder = NA, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{annotation_ids}{Annotation ID's for annotations to be included}

\item{modeloutput_ids}{Model Output ID's for model outputs to be included}

\item{buffer_before}{Number of seconds of audio to add before each annotation}

\item{buffer_after}{Number of seconds of audio to add after each annotation}

\item{buffer_btwn}{Number of seconds of silence to add between each annotation}

\item{sample_rate}{Sample rate of the merged recording}

\item{ammPath}{Path to the directory for the AMMonitor project (default NA)}

\item{storage_type}{What type of storage to use (defaults to "local")}

\item{media_folder}{Root file path or URL where the merged audio should be saved}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE.}
}
\value{
Returns 1 if successful
}
\description{
Create a merged audio file from specified 
annotations/modeloutputs to facilitate the verification workflow.
}
\details{
This function merges clips of user-specified annotations and model outputs 
into a single, merged audio file for easy verification.

The procedure for verifying audio annotations and model outputs often involves
reviewing a random sample of annotations/modeloutputs spread across many audio
files. This can create a bottleneck in the verification pipeline where a 
verifier must wait for an entire audio file to load in order to review just a 
small portion of it.

The user specifies the ID's of any annotations and/or model outputs that need
to be verified. A merged audio file will be created and associated with a 
special visit with visit_type "merged-verification". A mapping will be formed 
between annotations/modeloutputs in the merged audio and the annotations and 
modeloutputs from the original audio files. Once the merged audio file has 
been verified, the verifications from the merged audio file can be 
reconstituted into the database using the "mergedAudioRemove()" function.
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# Define input parameters (model outputs and annotations to be merged)
modeloutput_ids <- 7:9
annotation_ids <- 26:27

# Create a merged audio file and register in the database with the specified 
# annotations and model outputs
mergedAudioCreate(
  con = conx,
  annotation_ids = annotation_ids,
  modeloutput_ids = modeloutput_ids,
  ammPath = demo_fp
  )
  
  
# Look at the new visit that is associated with the merged file
(new_visit <- DBI::dbGetQuery(
  conx,
  "SELECT * FROM visits WHERE visit_type = 'merged-verification';"
  ))

# Look at the new, merged audio file; note the media type and filename
(merged_audio <- DBI::dbGetQuery(
  conx,
  paste("SELECT * FROM media WHERE fk_visitid = ", new_visit$pk_visitid, ";")
  ))  

# Verify the merged audio file exists in the recordings directory
file.exists(merged_audio$filepath)
list.files(file.path(demo_fp, "recordings"))

# Look at the annotations from the merged audio file
DBI::dbGetQuery(
  conx,
  paste(
    "SELECT * FROM annotations WHERE fk_mediaid = ",
    merged_audio$pk_mediaid,
    ";")
  )

# Look at the modeloutputs from the merged audio file
DBI::dbGetQuery(
  conx,
  paste(
    "SELECT * FROM modeloutputs WHERE fk_mediaid = ",
    merged_audio$pk_mediaid,
    ";")
  )
  
 # Look at the audio file in the shiny app; 
 # log in as unknownPerson
 # click on the Audio Tools tab
 # click on Annotation Verifications or Model Verifications to verify records 
 # in the merged audio file
 
 launchApp(demo_fp)
 
 # import records to the database with mergedAudioRemove()

# disconnect from the database when finished
DBI::dbDisconnect(conx)

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)
}
}
\seealso{
Other annotationverifications, modeloutputverifications: 
\code{\link{mergedAudioRemove}()}
}
\concept{annotationverifications, modeloutputverifications}
