% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedAudioRemove.R
\name{mergedAudioRemove}
\alias{mergedAudioRemove}
\title{Remove a merged audio file and associate verifications with the 
original annotations/modeloutputs}
\usage{
mergedAudioRemove(con, visitID = NA, mediaID = NA, ammPath = getwd(), 
disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database.}

\item{visitID}{the pk_visitid of the visit associated with the merged media 
to be removed. Default is NA. Use only one of visitID and mediaID.}

\item{mediaID}{the pk_mediaid of the merged media to be removed. Default is 
NA. Use only one of visitID and mediaID.}

\item{ammPath}{the path to the directory of the AMMonitor project. Default is 
getwd().}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed 
on exit? Default is FALSE.}
}
\description{
Removes a merged audio file created with mergedAudioCreate() 
from the database and associates all verifications with the original 
annotations and/or modeloutputs
}
\details{
This function takes a merged audio file created by 
"mergedAudioCreate()" and associates any verifications with the original
media file(s). The merged media file and any associated database entries will
then be deleted.

The user provides either a pk_visitid or pk_mediaid associated with the 
merged media file to be deleted. This function will then use the mapping 
formed by mergedAudioCreate() to assign any verifications to the original 
media. Once this process is complete, the function will remove all remaining 
references to the merged audio file from the database. All records will 
appear in the database as if they had been applied to the original audio 
files.
}
\examples{

\dontrun{

# Create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Set a connection to the demo database
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# Create a merged audio file from specified annotations and model verifications
modeloutput_ids <- 10:12
annotation_ids <- 26:27

mergedAudioCreate(
  con = conx,
  annotation_ids = annotation_ids,
  modeloutput_ids = modeloutput_ids,
  ammPath = demo_fp
  )

Look at the new visit that is associated with the merged file
DBI::dbGetQuery(
  conx,
  "SELECT * FROM visits WHERE visit_type = 'merged-verification';"
  )

# Confirm the new audio is present
list.files(file.path(demo_fp, "recordings"))
  
# Add some model verifications to the audio file as an example
# normally this would be done in the shiny app, but here coded for demonstration
DBI::dbAppendTable(
  conn = conx,
  name = "modelverifications",
  value = data.frame(
    pk_modelverificationid = NA,
    fk_modeloutputid = 56:58,
    fk_personid = "gandalf",
    is_valid = 1,
    timestamp = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S")
    )
  )

# Add annotation verifications to the audio file as an example
# normally this would be done in the shiny app, but here coded for demonstration
DBI::dbAppendTable(
  conn = conx,
  name = "annotationverifications",
  value = data.frame(
    pk_annoverificationid = NA,
    is_valid = 1,
    fk_personid = "gandalf",
    fk_annotationid = 28:29,
    timestamp = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S")
    )
  )   

# remove the merged media file and move verifications to the original media
mergedAudioRemove(
  con = conx,
  mediaID = 27,
  ammPath = demo_fp,
  disconnect = FALSE)
  
# Examine the updated modelverifications table
DBI::dbReadTable(
  conn = conx,
  name = "modelverifications")
  
# Examine the updated annotationverifications table
DBI::dbReadTable(
  conn = conx,
  name = "annotationverifications")

# disconnect from the database when finished
DBI::dbDisconnect(conx)

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)
}
}
\seealso{
Other annotationverifications, modeloutputverifications: 
\code{\link{mergedAudioCreate}()}
}
\concept{annotationverifications, modeloutputverifications}
