% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAdd.R
\name{modelAdd}
\alias{modelAdd}
\title{Adds a model, settings, and options to an AMMonitor database}
\usage{
modelAdd(con, excel_fp, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{excel_fp}{A filepath to an Excel file containing new model information.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
An updated database with new model info
}
\description{
This function adds a model, along with its settings and any
associated options, to an AMMonitor database using an Excel spreadsheet
provided with this package. See Details for more information.
}
\details{
The modelAdd() function simplifies the process of adding models to your
database by allowing you to add a new model, its associated settings, and any
setting options in a single function. This approach can be useful for adding
models programmatically, or for easily adding a single model to multiple
projects. Models can also be added to the database using the modelManager
app; full documentation of this app can be found in the "models" tutorial.

This function relies on a template Excel file that is provided with this 
package. You can copy the template to a directory of your choice with the 
following code:

\code{fp <- system.file("extdata/modelAdd.xlsx", package = "AMMonitor")}
 
\code{file.copy(from = fp, to = tempdir(), overwrite = FALSE)}

Open this Excel file and fill in the provided fields for the model you'd like
to add. Do not alter the original template; AMMonitor relies on this format
to read in the provided model info properly. Each table in the spreadsheet 
corresponds to a table in the database:

Table 1: Model (Blue) corresponds to the models table.

Table 2: Settings (Green) corresponds to the msettingnames table.

Table 3: Options (Orange) corresponds to the msettingoptions table.

The fields in the spreadsheet correspond to the database tables by name. For
more information AMMonitor, see the "models" tutorial, which can be launched 
with the following code:
\code{learnr::run_tutorial(name = "models", package = "AMMonitor")}
}
\examples{

\dontrun{
# Create a demo database
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Connect to the database
conx <- dbSetCon(file.path(demo_fp, "/database/demo.sqlite"))

# Copy the template spreadsheet to your demo directory
fp <- system.file("extdata/modelAdd.xlsx", package = "AMMonitor")
file.copy(from = fp, to = demo_fp, overwrite = FALSE)

# Fill in the tables with example model information

# Add the model to the database
modelAdd(
  con = conx,
  excel_fp = file.path(demo_fp, "modelAdd.xlsx"),
  disconnect = FALSE)
  
# Check the updated models table
DBI::dbReadTable(conx, "models")

# Check the updated msettingnames table
DBI::dbReadTable(conx, "msettingnames")

# Check the updated msettingoptions table
DBI::dbReadTable(conx, "msettingoptions")

# remove the excel template
unlink(file.path(tempdir(), "modelAdd.xlsx"))

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)
}

}
