% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelAddConfig.R
\name{modelAddConfig}
\alias{modelAddConfig}
\title{Adds a model configuration to an AMMonitor database}
\usage{
modelAddConfig(con, excel_fp, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{excel_fp}{A filepath to an Excel file containing new model 
configuration information.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
An updated database containing a new model configuration
}
\description{
This function adds a model configuration to an AMMonitor
database and updates all associated tables. See details for further info.
}
\details{
The modelAddConfig() function is designed to enable easy addition of new
model configurations to an AMMonitor database using an Excel spreadsheet. You
can generate an Excel spreadsheet for a specific model using the
modelConfigWriteXL() function. 

If the model hasn't been added to the database yet, see the modelAdd() 
function. Models and configurations can also be added to the database using
the modelManager app in AMMonitor; see the "models" tutorial for more
information on both methods.

To add a configuration via spreadsheet, open up the newly generated 
spreadsheet and fill out the fields to create a new model configuration. Each
of the three sheets corresponds to different database tables:

Configuration_Name corresponds to the mconfignames table and associated
fields. The model ID is automatically filled in when the spreadsheet is
generated; other fields should be filled in.

Configuration_Values corresponds to the mconfigvalues table and provides the
names and descriptions of all valid settings for the specified model, as well
as Option_Name and Value columns to choose settings. Fill in the Option_Name
and Value columns, and leave a blank cell for empty entries.

Setting_Options corresponds to the msettingnames and msettingoptions tables.
It provides a list of all valid options for settings with options in the
database. This sheet does not need to be updated, but can be used for
reference when selecting setting options.

You can find more information about model configurations and the associated
database tables in the "models" tutorial, including a full walkthrough of
adding new configurations. You can launch this tutorial as follows:
\code{learnr::run_tutorial(name = "models", package = "AMMonitor")}
}
\examples{
\dontrun{
# Create a demo database
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Connect to the database
conx <- dbSetCon(file.path(demo_fp, "/database/demo.sqlite"))

# Take a look at the models table
DBI::dbReadTable(conx, "models")

# Let's make a spreadsheet for model 6, BirdNET, and write it to the demo dir
modelConfigWriteXL(
  con = conx,
  modelid = 6,
  output_dir = demo_fp,
  disconnect = FALSE)

# Fill in the spreadsheet with example model configuration info

# Add your new configuration to the database
modelAddConfig(
  con = conx,
  excel_fp = file.path(demo_fp, "modelAddConfig.xlsx")),
  disconnect = FALSE)
  
# Run a query to see your new model configuration
qryMconfiguration(
  con = conx, 
  mconfigid = 3,
  disconnect = FALSE)
}

}
