% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelConfigWriteXL.R
\name{modelConfigWriteXL}
\alias{modelConfigWriteXL}
\title{Writes an Excel spreadsheet to be used to easily add a new model
configuration to an AMMonitor database}
\usage{
modelConfigWriteXL(con, modelid, output_dir, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{modelid}{The pk_modelid of the model you'd like to add a configuration
for, as listed in the "models" table}

\item{output_dir}{A filepath to the directory where you'd like the output
Excel spreadsheet to be saved.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
An Excel spreadsheet containing tables to add a new model 
configuration to the database in the specified directory
}
\description{
This function writes an Excel spreadsheet to a given directory
with sheets to add a new model configuration to an AMMonitor database. Each
sheet corresponds to a different database table. For more information, see 
the Details section.
}
\details{
This function creates an Excel spreadsheet to enable easy addition of model
configuration to an AMMonitor database. It queries the database based on the
provided model ID to find all settings and setting options for the given
model and provides an organized spreadsheet with relevant entires to create
a new model configuration.

The resulting spreadsheet will be named "modelAddConfig.xlsx" and written to
the provided output directory. This spreadsheet contains three sheets:

Configuration_Name corresponds to the mconfignames table and associated
fields. The model ID is automatically filled in based on the provided modelid
in the function.

Configuration_Values corresponds to the mconfigvalues table and provides the
names and descriptions of all valid settings for the specified model, as well
as Option_Name and Value columns to choose settings.

Setting_Options corresponds to the msettingnames and msettingoptions tables.
It provides a list of all valid options for settings with options in the
database. An empty row next to a setting indicates that the setting takes a
numeric input in the Value column.

Model configurations can also be added using the modelManager app in 
AMMonitor. For details on using the app to add configurations, see the
"models" tutorial.

For details on adding the configuration to the database via this spreadsheet,
see modelAddConfig(). For a full walkthrough, see the "models" tutorial,
which can be launched using the following code:
\code{learnr::run_tutorial(name = "models", package = "AMMonitor")}
}
\examples{
\dontrun{
# Create a demo database
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Connect to the database
conx <- dbSetCon(file.path(demo_fp, "/database/demo.sqlite"))

# Take a look at the models table
DBI::dbReadTable(conx, "models")

# Let's make a spreadsheet for model 6, BirdNET, and write it to the demo dir
modelConfigWriteXL(
  con = conx,
  modelid = 6,
  output_dir = demo_fp,
  disconnect = FALSE)
}

}
