% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVerifications.R
\name{plotVerifications}
\alias{plotVerifications}
\title{Plot side-by-side spectrograms of verified data}
\usage{
plotVerifications(
  con,
  model_id,
  taxon_id,
  plot_modeloutput_id = FALSE,
  min_value_num = 0,
  limit = 10,
  amm_project_path = getwd(),
  spec_wl = 512,
  spec_wn = "hanning",
  spec_zp = 0,
  spec_ovlp = 0,
  spec_freq_range = c(0, 10),
  palette = "reverse.gray.colors.1",
  disconnect = FALSE
)
}
\arguments{
\item{con}{An open database connection}

\item{model_id}{pk_modelid.}

\item{taxon_id}{A valid primary key reference to the taxa table.}

\item{plot_modeloutput_id}{Whether to plot the modeloutputid on the spectrogram}

\item{min_value_num}{Mimimum value_num for filtering modeloutputs.}

\item{limit}{The maximum number of valid and invalid examples to plot (default = 10).}

\item{amm_project_path}{Filepath to the outermost directory for an AMMonitor project (defaults to current working directory)}

\item{spec_wl}{window length for the analysis (even number of points) (by default = 512).}

\item{spec_wn}{window name (by default "hanning").}

\item{spec_zp}{zero-padding (even number of points)}

\item{spec_ovlp}{overlap between two successive windows (in percent).}

\item{spec_freq_range}{Frequency range, in KHz (defaults to c(0, 10))}

\item{palette}{Seewave color palette for plotting spectorgrams (defaults to "reverse.gray.colors.1")}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed}
}
\value{
A ggplot object with two columns; the left containing valid (verified true) modeloutputs for the given taxon, and the right containing invalid (verified false) detections. \code{plotVerifications} also returns the pk_modeloutputid of each plotted verification from the modeloutputs table.
}
\description{
Select a model, taxon, and score threshold, and compare validated modeloutputs (audio only) based on whether they were maked as valid or invalid. Side-by-side spectrograms are plotted so the user can make a visual comparison. The modeloutput ID's are returned.
}
\details{
Select a model, taxon, and score threshold, and compare validated modeloutputs based on whether they were maked as valid or invalid. Side-by-side spectrograms are plotted so the user can make a visual comparison. The modeloutput ID's are returned.
}
\examples{
\dontrun{

# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# Add some additional BirdNET verifications (let's make this more interesting)
new_modelverifications <- data.frame(
  fk_modeloutputid = c(45, 52, 46),
  fk_personid = c("sgamgee", "sgamgee", "sgamgee"),
  is_valid = c(1, 0, 1)
)
DBI::dbAppendTable(conx, "modelverifications", new_modelverifications)

# Plot the verifications (plot will be displayed)
modeloutput_ids <- AMMonitor::plotVerifications(
  con = conx, 
  model_id = 6, 
  taxon_id = "oven", 
  plot_modeloutput_id = TRUE, 
  amm_project_path = demo_fp
)

# View modeloutput_id's from the returned sample of verificatied modeloutputs
(modeloutput_ids)

# disconnect from the database when finished
DBI::dbDisconnect(conx)

# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)

}
}
\seealso{
Other classifier: 
\code{\link{testThresholdDetx}()},
\code{\link{testThresholdRec}()}
}
\concept{classifier}
