% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qry_row}
\alias{qry_row}
\title{Retrieve one or more rows from a specified table from the AMMonitor
database given a set of conditions on one or more columns.}
\usage{
qry_row(con, tableName, rowConditions = NA, colConditions = "*", disconnect = FALSE)
}
\arguments{
\item{tableName}{The name of the table of the AMMonitor database (e.g. 
people, media, locations)}

\item{rowConditions}{A dataframe where the keys correspond to columns of the
specified dataframe and key values correspond to the equality condition that
must be satisfied by any returning rows, else returns all rows 
(default returns all rows).}

\item{colConditions}{A vector specifying the names of columns to be returned
from the query (default returns all columns).}

\item{con}{Existing connection to the database}

\item{disconnect}{TRUE or FALSE. Should the connection be severed on exit?
Default is FALSE.}
}
\value{
Dataframe consisting of the specified rows and columns
}
\description{
Returns dataframe containing specified columns and rows from the
AMMonitor database based on specified conditions.
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

qry_row(
  con = conx,
  tableName = 'locations',
  rowConditions = list(location_type = 'monitoring_station'),
  colConditions = 'pk_locationid',
  disconnect = TRUE
)

# unlink the demo
unlink(demo_fp)

}
}
\seealso{
Other qry: 
\code{\link{qryAnnotations}()}
}
\concept{qry}
