% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryAnnotations}
\alias{qryAnnotations}
\title{Returns a count of annotations by taxa in the database.}
\usage{
qryAnnotations(con,  disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{disconnect}{TRUE or FALSE. Should the connection to the database be 
closed on exit? Default is FALSE.}
}
\value{
Dataframe consisting of the specified rows and columns
}
\description{
Returns a count of annotations by taxa in the database and the 
number of records verified
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

rslt <- qryAnnotations(
  con = conx,
  disconnect = TRUE)

# show the result
rslt

# make a table with the package gt
gt::gt(rslt) |> 
  gt::cols_label(
    fk_taxonid = "Taxa",
    total_annotations = "Annotations",
    total_verified = "Verifications") 
 
# unlink the demo
unlink(demo_fp)




# unlink the demo
unlink(demo_fp)
}

}
\seealso{
Other qry: 
\code{\link{qry_row}()}
}
\concept{annotation}
\concept{annotationverifications}
\concept{qry}
