% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryCheckMediaFilePaths}
\alias{qryCheckMediaFilePaths}
\title{Check if any row in the media table does not have a file path}
\usage{
qryCheckMediaFilePaths(con, mediaType = NA, mediaIDs = NA, 
disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{mediaType}{Media type to filter for, if any. One of "photo", "audio", or 
"video". If mediaIDs are specified, this argument will be ignored.}

\item{mediaIDs}{Integer vector of media IDs from the media table (pk_mediaid), 
if specific ones should be selected.}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Checks if there exists any row in the media table that meets 
certain conditions (media type or media ID) and has no file path
}
\details{
File paths are not required in the media table! This function is used 
internally to determine if file paths are provided, and if not, the Shiny app
will look for the media root file path in the "settings" directory. If not 
present the app will assume the media are present in the "photos" or 
"recordings" directory of any AMMonitor project.
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the models table; note that filepaths are not present
DBI::dbReadTable(conx, name = "media")

# check if filepaths are provided
qryCheckMediaFilePaths(conx, mediaType = "photo", disconnect = TRUE)

# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{media}
\concept{query}
