% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryCheckTags}
\alias{qryCheckTags}
\title{Query annotations or mediatags for specific media files}
\usage{
qryCheckTags(
  con, 
  mediaid, 
  personid = NA, 
  excludeperson = FALSE, 
  mediatags = FALSE, 
  exists = FALSE, 
  disconnect = FALSE
)
}
\arguments{
\item{con}{An existing database connection}

\item{mediaid}{One or more mediaids as an integer vector, as saved in the database.}

\item{personid}{A personid as saved in the database, if specified, then
filters for or out annotations by the specified user. Default NA}

\item{excludeperson}{TRUE or FALSE. Whether to include or exclude annotations
by the specified user. Default is FALSE}

\item{mediatags}{TRUE or FALSE. Should the function check for mediatags 
instead of annotations? Default is FALSE}

\item{exists}{TRUE or FALSE. Whether to check for existing records or to 
retrieve those records. Default is FALSE}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Check if annotations or mediatags exist for a particular media file
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the media table; note that filepaths are not present
DBI::dbReadTable(conx, name = "media")

# return all library tags for media 1:5
qryCheckTags(conx, mediaid = 1:5, disconnect = FALSE)

# look at the mediatags table; there are none
DBI::dbReadTable(conx, name = "mediatags")

# look at the media tag items available
qryItems(
  con = conx, 
  table = "medialists", 
  listname = "media_checkboxes"
 )

# add a few media-level tags for media 1:5
DBI::dbAppendTable(
 conn = conx,
 name = "mediatags",
 value = data.frame(
  pk_mediatagid = NA,
  fk_mediaid = 1:5,
  fk_medialistitemid = 1,  # highlight
  fk_personid = 'gandalf'
  )
 )

# check all media tags for media 1:5
qryCheckTags(conx, mediaid = 1:5, mediatags = TRUE, disconnect = TRUE)

# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{annotations}
\concept{media}
\concept{mediatags}
\concept{query}
