% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryEconfiguration}
\alias{qryEconfiguration}
\alias{qryEconfigurations}
\title{Returns settings and values for a given equipment configuration}
\usage{
qryEconfiguration(con, econfigname, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{econfigname}{The primary key in the econfignames table}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Returns settings and values for a given equipment configuration
}
\examples{
\dontrun{
# create a demo AMMonitor project (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# look at the econfigs table
DBI::dbReadTable(conx, name = "econfignames")

# Retrieve a list of settings for a given equipment configuration
qryEconfiguration(
 con = conx,
 econfigname = "default recorder settings",
 disconnect = TRUE)
 
# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{econfignames}
\concept{econfigvalues}
\concept{equipment}
\concept{equipmodels}
\concept{esettingnames}
\concept{esettingoptions}
\concept{query}
