% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryEffort}
\alias{qryEffort}
\title{Returns dates used to estimate survey effort per visit}
\usage{
qryEffort(con, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
List containing two dataframes
}
\description{
Returns lower-bound (based on media dates) and upper-bound
 (based on visit dates and equipment deployment) date ranges used to 
 calculate survey effort per visit. Visits with 'NA' in the lower-bound
  columns were not associated with collected media. Visits with 'NA' 
  in the upper-bound start date 'activeStartUB' have no prior visit in 
  the database.
}
\details{
This function provides estimates of survey effort with upper and lower bounds.
Lower bound estimates for each visit were calculated from the minimum and 
maximum media dates for that visit. The upper bound estimates for each visit 
were calculated as the difference between a visit's date and the date of the 
previous visit at that location.
}
\examples{
\dontrun{
# create a demo AMMonitor project (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# retrieve estimates of survey effort
effort <- qryEffort(con = conx, disconnect = TRUE)

# look at the dates estimating survey effort per visit
effort
 
# unlink the demo
unlink(demo_fp)
}

}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{query}
\concept{visits}
