% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryEquipModelSettings}
\alias{qryEquipModelSettings}
\title{Queries the database and returns a data frame including all settings
and setting options for a given equipment model}
\usage{
qryEquipModelSettings(con, equipmodelid, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{equipmodelid}{A pk_equipmodelid, as listed in the "equipmodels" table 
of the database}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
A dataframe containing all valid settings and setting options for
the provided equipmodel ID
}
\description{
This query provides a full list of settings and setting options
associated with a given equipmodel ID.
}
\examples{
\dontrun{
# Create a demo database
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Connect to the database
conx <- dbSetCon(file.path(demo_fp, "/database/demo.sqlite"))

# Look at the equipmodels table
DBI::dbReadTable(conx, "equipmodels")

# Query settings for recorder_model_x
qry <- qryEquipModelSettings(
 con = conx,
 equipmodelid = "recorder_model_x",
 disconnect = FALSE)
 
# View the query results
View(qry)
}

}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{equipmodels}
\concept{esettingnames}
\concept{esettingoptions}
\concept{query}
