% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryItems}
\alias{qryItems}
\title{Retrieves all items from a given list}
\usage{
qryItems(con, table, listname, item = NA, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{table}{The database table name. Options are "lists", "medialists", 
"librarylists" or "temporallists"}

\item{listname}{The name of a list}

\item{item}{The name of a particular item with in a list, if desired.}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Retrieve items associated with a list, a librarylist, or a medialist.
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# retrieve items from the listitems table
qryItems(
  con = conx, 
  table = "lists", 
  listname = "datum", 
  item = "WGS84",
  disconnect = FALSE
)

# retrieve all items from the media_type list
qryItems(
  con = conx, 
  table = "lists", 
  listname = "media_type",
  disconnect = TRUE
 )
 
# unlink the demo
unlink(demo_fp)

}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{librarylistitems}
\concept{librarylists}
\concept{listitems}
\concept{lists}
\concept{medialistitems}
\concept{medialists}
\concept{query}
