% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryLocationCounts}
\alias{qryLocationCounts}
\title{Returns counts of locations by type as 
logged in the database}
\usage{
qryLocationCounts(con, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Returns counts of locations by type as 
logged in the database
}
\examples{
\dontrun{
# create a demo AMMonitor project (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# retrieve a count of locations by type
qryLocationCounts(
 con = conx,
 disconnect = TRUE)
 
# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{locations}
\concept{query}
