% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryMconfiguration}
\alias{qryMconfiguration}
\title{Returns settings and values for a given model configuration}
\usage{
qryMconfiguration(con, mconfigid, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{mconfigid}{The pk_mconfigid from the mconfignames table}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Returns settings and values for a given model configuration
}
\details{
This query retrieves information about model configurations and presents it
in an easy-to-read format. By default, this information will be output in a
data frame.
}
\examples{
\dontrun{
# create a demo AMMonitor project (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# look at the mconfignames table
# note that the BirdNET default settings are pk_mconfigid 2
DBI::dbReadTable(conx, name = "mconfignames")

# Retrieve the settings and values for the default BirdNET configuration
qryMconfiguration(
 con = conx,
 mconfigid = 2,
 disconnect = TRUE)
 
# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{mconfignames}
\concept{mconfigvalues}
\concept{models}
\concept{msettingnames}
\concept{msettingoptions}
\concept{query}
