% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryMedia}
\alias{qryMedia}
\title{Filter a list of media based on many criteria}
\usage{
qryMedia(con, 
 disconnect = FALSE, 
 locationID = 'all',
 dateRange = list(c('1900-01-01', as.character(Sys.Date()))), 
 visitID = NULL, 
 taxonID = 'all', 
 excludeAnnotated = 'NA', 
 excludeAnnoVerified = 'NA', 
 showAnnotated = FALSE, 
 showUNAnnotated = FALSE, 
 selectedUser = NA, 
 verify = FALSE, 
 mediaType = "photo",
 limit = Inf,
 offset = 0
)
}
\arguments{
\item{con}{SQLite database connection}

\item{disconnect}{TRUE or FALSE Should the database connection be closed on 
exit? Default is FALSE.}

\item{locationID}{specify a location (pk_locationid)}

\item{dateRange}{Specify a date range}

\item{visitID}{Specify a specific visit (pk_visitid)}

\item{taxonID}{Specify a specific taxon (pk_taxonid)}

\item{excludeAnnotated}{Specify when to exclude based on annotation status}

\item{excludeAnnoVerified}{Specify when to exclude based on verification status}

\item{showAnnotated}{TRUE or FALSE. Show only annotated files. 
Default is FALSE}

\item{showUNAnnotated}{TRUE or FALSE. Show only un-annotated files. 
Default is FALSE}

\item{selectedUser}{The selected user (if any) to filter by taxon 
(pk_personid)}

\item{verify}{TRUE or FALSE. If tagger is in "verify" mode. Default is FALSE}

\item{mediaType}{type of media to filter for, one of (photo, audio, video).
Default is "photo."}

\item{limit}{Maximum number of records to be returned}

\item{offset}{The number of records to offset the query by}
}
\value{
A dataframe with key columns for all media that meet the given criteria
}
\description{
Filter a list of media based on many criteria
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# run the query to get media that have moose annotations
qryMedia(conx, taxonID = "moose", disconnect = FALSE)

# run a location query to get media associated with locationA
qryMedia(conx, locationID = "locationA", disconnect = FALSE)

# run a query to get media associate with visit 1
qryMedia(conx, visitID = 10, disconnect = TRUE)

# unlink the demo
unlink(demo_fp)
}

}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{annotations}
\concept{annotationverification}
\concept{locations}
\concept{media}
\concept{query}
\concept{taxa}
\concept{visits}
