% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryMediaDateRange}
\alias{qryMediaDateRange}
\title{Queries the date range of all media files or a type of media files}
\usage{
qryMediaDateRange(con, mediaType = NA, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{mediaType}{Type of media to filter for, if any. One of "photo", "audio"
or "video". Default is NA (all media).}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on 
exit? Default is FALSE.}
}
\value{
dataframe with the first column being the start date and the second column being the end date
}
\description{
Queries the date range of all media files or a type of media files
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the media table
DBI::dbReadTable(conx, name = "media")

# return the date range of photos
qryMediaDateRange(conx, mediaType = "photo", disconnect = FALSE)

# return the date range of photos
qryMediaDateRange(conx, mediaType = "audio", disconnect = TRUE)

# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{media}
\concept{query}
