% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryMediaLocations}
\alias{qryMediaLocations}
\title{Retrieves locations with media files of a certain type}
\usage{
qryMediaLocations(con, mediaType, disconnect)
}
\arguments{
\item{con}{An existing database connection}

\item{mediaType}{Type of media, one of "photo", "audio", "video"}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed 
on exit? Default is FALSE.}
}
\value{
character vector of all available locations
}
\description{
Retrieves distinct locations with specified types of media files
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# get the locations with photos
qryMediaLocations(con = conx, mediaType = "photo", disconnect = FALSE)

# get the locations with recordings
qryMediaLocations(con = conx, mediaType = "audio", disconnect = TRUE)

# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{locations}
\concept{media}
\concept{query}
\concept{visits}
