% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryModelOutputs}
\alias{qryModelOutputs}
\title{Queries the model outputs table, with filters by score, model, and media 
file}
\usage{
qryModelOutputs(con, mediaIDs = NA, modelIDs = NA, scoreThreshold = NA,
lessThan = FALSE, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{mediaIDs}{Integer vector - One or more media IDs (as saved in the database) 
to get model outputs for}

\item{modelIDs}{Integer vector - One or more model IDs (as saved in the database) 
to get model outputs for}

\item{scoreThreshold}{Numeric - Threshold of scores to filter for}

\item{lessThan}{TRUE or FALSE. Whether or not retrieved outputs should have a 
score greater than (default) or less than the threshold. Only applicable if 
scoreThreshold is specified. Default is FALSE, which means that outputs 
retrieved will have scores greater than the threshold.}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
dataframe with the first column being the start date and the second 
column being the end date
}
\description{
Queries the model outputs table, with filters by score, model, and 
media file
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

  
# return modeloutputs from media 10
qryModelOutputs(
  con = conx, 
  mediaIDs = 10:12, 
  modelIDs = NA, 
  scoreThreshold = NA, 
  lessThan = TRUE, 
  disconnect = FALSE)
  
# return modeloutputs from all media, model 5
qryModelOutputs(
  con = conx, 
  mediaIDs = NA, 
  modelIDs = 5, 
  scoreThreshold = NA, 
  lessThan = TRUE, 
  disconnect = FALSE)
  
# return modeloutputs from media 10-12
qryModelOutputs(
  con = conx, 
  mediaIDs = 10:12, 
  modelIDs = 5:6, 
  scoreThreshold = NA, 
  lessThan = TRUE, 
  disconnect = FALSE)
  
# return modeloutputs from media 10-12 with high score threshold
qryModelOutputs(
  con = conx, 
  mediaIDs = 10:12, 
  modelIDs = 5:6, 
  scoreThreshold = .95, 
  lessThan = TRUE, 
  disconnect = FALSE)

# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{modeloutputs}
\concept{query}
