% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryModelOutputsFull}
\alias{qryModelOutputsFull}
\title{Returns a data frame of model outputs that meet a variety of criteria}
\usage{
qryModelOutputsFull(
   con, 
   mediaIDs = NA,
   modelIDs = NA,
   locationIDs = NA,
   dateRange = list(c("1900-01-01", as.character(Sys.Date()))),
   taxonIDs = NA,
   value_num = NA,
   lessThan = FALSE,
   disconnect = FALSE
   )
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{mediaIDs}{A vector of pk_mediaids to get model outputs for. Default is 
NA.}

\item{modelIDs}{A vector of pk_modelids to get model outputs for. Default is
NA.}

\item{locationIDs}{A vector of pk_locationids to get model outputs for. 
Default is NA.}

\item{dateRange}{A range of dates to get model outputs from. Dates should be
formatted as character data and placed in a list. Default is 1900-01-01 to 
the current date.}

\item{taxonIDs}{A vector of pk_taxonids to get model outputs for. Default is
NA.}

\item{value_num}{A value to filter model outputs by, based on the value_num
in the model outputs table (such as a confidence score). Default is NA.}

\item{lessThan}{TRUE or FALSE. Should outputs be less than the specified
value_num? Default is FALSE.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
A data.frame of all model outputs which match the provided criteria
}
\description{
This function allows for complex filtering of model outputs to
suit various needs. Filtering options include date range, location, taxa,
value_num, model, and media.
}
\details{
This function builds upon the functionality of qryModelOutputs to allow for
more complex filtering of model outputs. All arguments except for dateRange
can be set to NA to ignore them; in the case of dateRange, the default will
include all possible model outputs. All IDs should be entered as they are 
listed in the corresponding primary key (pk) in the database. The returned
data frame will include all key information from the modeloutputs table for
matching outputs, as well as the start_date and location for the output.
}
\examples{
\dontrun{

# Create a demo AMMonitor project (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# To work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# Only get modeloutputs for Ovenbird (oven)
qryModelOutputsFull(
   con = conx,
   taxonIDs = "oven",
   disconnect = FALSE)

# Only get modeloutputs from BirdNET (modelid = 6)
qryModelOutputsFull(
   con = conx,
   modelIDs = 6,
   disconnect = FALSE)
}

}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{modeloutputs}
\concept{query}
