% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryModelSettings}
\alias{qryModelSettings}
\title{Queries the database and returns a data frame including all settings
and setting options for a given model}
\usage{
qryModelSettings(con, modelid, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{modelid}{A pk_modelid, as listed in the "models" table of the database}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
A dataframe containing all valid settings and setting options for
the provided model ID
}
\description{
This query provides a full list of settings and setting options
associated with a given model ID.
}
\examples{
\dontrun{
# Create a demo database
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Connect to the database
conx <- dbSetCon(file.path(demo_fp, "/database/demo.sqlite"))

# Look at the models table
DBI::dbReadTable(conx, "models")

# Query settings for model 6, BirdNET
qry <- qryModelSettings(
 con = conx,
 modelid = 6,
 disconnect = FALSE)
 
# View the query results
View(qry)
}

}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{models}
\concept{msettingnames}
\concept{msettingoptions}
\concept{query}
