% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryModelVerificationConsensus}
\alias{qryModelVerificationConsensus}
\title{Retrieves model outputs of a particular model with a consensus in
verifications -- all people who verified agree on 1 or 0 (even if there is 
just one person making the verifications)}
\usage{
qryModelVerificationConsensus(con, modelid, disconnect = FALSE)
}
\arguments{
\item{con}{An open database connection}

\item{modelid}{Model ID of a model (integer, pk_modelid from the models table)}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on 
exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Retrieves model outputs of a particular model with a consensus 
in verifications (for >= 1 person).  If there are conflicts (e.g., one
person confirms valid and another confirms invalid, these records are 
excluded)
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the models table; note that MegaDetector is model 5
DBI::dbReadTable(conx, name = "models")

# look at the modeloutputs table
(mo <- DBI::dbReadTable(conx, name = "modeloutputs"))

# look at the modelverifications table; note sgamgee made verifications
# on multiple model outputs, some of them are MegaDetector outputs
(mvs <- DBI::dbReadTable(conx, name = "modelverifications"))

# add Gandalf's modelverifications as well - but set these to random values!!
# (Gandalf should verify more carefully)
DBI::dbAppendTable(
  conn = conx,
  name = "modelverifications",
  value = data.frame(
   pk_modelverificationid = NA,
   fk_modeloutputid = mo$pk_modeloutputid,
   fk_personid = "gandalf",
   is_valid = stats::rbinom(n = nrow(mo), size = 1, prob = 0.5),
   timestamp =format(Sys.time())
  )
 )

# run the query to get MegaDetector modeloutputs that have been validated by 
# at least two people; note these contain only MegaDetector outputs (model 5)
result <- qryModelVerificationConsensus(conx, modelid = 5, disconnect = TRUE)

# graph the consensus
ggplot2::ggplot(result, aes(x = as.factor(is_valid))) +
  geom_bar() + 
  labs(
    x = "Is Valid",
    y = "Count")


# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{modeloutputs}
\concept{models}
\concept{modelverifications}
\concept{query}
