% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryPeopleActivity}
\alias{qryPeopleActivity}
\title{Returns information about activities by person as 
logged in the database}
\usage{
qryPeopleActivity(con, m_dates = c(NA, NA), disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{m_dates}{A vector of length 2 giving start and end date ranges 
for report, in 
the format YYYY-MM-DD. Default is c(NA, NA).}

\item{disconnect}{TRUE or FALSE.  Should the database connection be closed 
on exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Returns information about activities as by person
logged in the database
}
\examples{
\dontrun{
# create a demo AMMonitor project (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# Retrieve a full report for a given person
results <- qryPeopleActivity(
 con = conx, 
 m_dates = c(NA, NA),
 disconnect = TRUE)
 
 # look at the results
 lapply(results, FUN = head, n = 10)
 
ggplot2::ggplot(
 data = results$visits,
 mapping = aes(
   fill = as.factor(visit_type), 
   x = fk_personid)) +
 geom_bar() +
 labs(
   x = "Person",
   y = "Count",
   fill = "Visit Type") +
 scale_y_continuous(breaks=c(1:9))
 
# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{annotations}
\concept{annotationverifications}
\concept{modelverifications}
\concept{people}
\concept{query}
\concept{visits}
