% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryTags}
\alias{qryTags}
\title{Retrieves all tags that can be applied to a media file}
\usage{
qryTags(con, media = FALSE, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{media}{TRUE or FALSE. Should tags associated with full media be 
returned? Default is FALSE.}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit?  Default is FALSE.}
}
\value{
dataframe
}
\description{
Retrieves all tags that can be applied to a media file. (These
are not the actual tags; rather they are the possible tags.)
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# Retrieve library tags only (tags that can be applied to taxa)
qryTags(con = conx, media = FALSE, disconnect = FALSE)

# Retrieve library and media tags
qryTags(con = conx, media = TRUE, disconnect = TRUE)

# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{librarylistitems}
\concept{librarylists}
\concept{medialistitems}
\concept{medialists}
\concept{query}
