% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryUnannotated}
\alias{qryUnannotated}
\title{Returns files without annotations}
\usage{
qryUnannotated(con, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Returns files without annotations
}
\examples{
\dontrun{
# create a demo AMMonitor project (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# look at the "annotations" table.
# note that all of the media in the demo have annotations
DBI::dbReadTable(conn = conx, name = "annotations")

# delete some of the example annotations
DBI::dbExecute(
 conn = conx,
 statement = "DELETE FROM annotations
 WHERE fk_mediaid = 1;")
 
# run the query to see a list of files lacking annotations
qryUnannotated(con = conx, disconnect = TRUE)
 
# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{annotations}
\concept{media}
\concept{query}
\concept{visits}
