% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryVerifications}
\alias{qryVerifications}
\title{Queries the annotations, mediatags, or model outputs tables and 
respective verifications}
\usage{
qryVerifications(con, table, personid = NA, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{table}{Type of verifications to query, one of "annotations", "mediatags", 
or "modeloutputs".}

\item{personid}{personid as saved in the database. If not specified, then 
verifications by all users will be retrieved.}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on 
exit? Default is FALSE.}
}
\value{
dataframe with the first column being the start date and the second 
column being the end date
}
\description{
Queries the annotations, mediatags, or model outputs tables and 
their respective verifications tables. Verifications can be filtered by 
database user
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# return annotation verifications
qryVerifications(conx, table = 'annotations', disconnect = FALSE)

# return modeloutputs verifications
rslt <- qryVerifications(conx, table = 'modeloutputs', disconnect = FALSE)

# graph the result
ggplot2::ggplot(
 data = rslt[,1:11],
 mapping = aes(
   fill = as.factor(is_valid), 
   x = fk_taxonid)) +
 geom_bar() +
 labs(
   x = "Taxa",
   y = "Count",
   fill = "Valid") 

# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVisitTable}()}
}
\concept{annotations}
\concept{annotationverifications}
\concept{mediatags}
\concept{mediatagverifications}
\concept{modeloutputs}
\concept{modelverifications}
\concept{query}
