% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryVisitTable}
\alias{qryVisitTable}
\title{Retrieves counts of files and annotated files per visit to a location}
\usage{
qryVisitTable(con, mediaType, location, personid, disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection}

\item{mediaType}{type of media to filter for. Options are "audio" or "photo".}

\item{location}{location to retrieve visits from}

\item{personid}{personid to retrieve annotations from}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on
exit? Default is FALSE.}
}
\value{
dataframe
}
\description{
Retrieves counts of files and annotated files per visit to a location
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# get information on photo media files collected at locationA 
qryVisitTable(
 con = conx, 
 mediaType = "photo", 
 location = "locationA", 
 personid = NA, 
 disconnect = FALSE)
 
# get information on audio media files collected at locationA 
qryVisitTable(
 con = conx, 
 mediaType = "audio", 
 location = "locationA", 
 personid = NA, 
 disconnect = TRUE)

# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelOutputsMedia}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()}
}
\concept{locations}
\concept{media}
\concept{people}
\concept{query}
\concept{visits}
