% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s123Create.R
\name{s123Create}
\alias{s123Create}
\title{Creates a Survey123 spreadsheet from an AMMonitor database}
\usage{
s123Create(con, equip_type = 'all',
 reminders = NULL, notes = NULL, disconnect = FALSE)
}
\arguments{
\item{con}{Open connection to an AMMonitor database}

\item{equip_type}{Type of equipment to include from 
the database. Options are 'camera', 'recorder', default is 'all'. 
Any equipment in the database with equip_type = NA will be included.}

\item{reminders}{An optional vector of 'checkbox'
reminders to add to the end of the survey form}

\item{notes}{An optional vector of notes to add to 
the end of the survey form}

\item{disconnect}{TRUE or FALSE. Should the 
database connection be severed on exit? 
Default is FALSE}
}
\value{
filepath to the data source .xlsx file
}
\description{
Populates the Survey123 Excel 
spreadsheet to enable users to collect visit data from a Survey123
cell phone app. The function assumes there are entries in the people, 
locations, and equipment tables of the database.
 
See the 'mobile_apps' learnr tutorial for instructions on creating the survey 
in  ArcGIS Survey123 Connect.
}
\details{
This function creates an Excel workbook from a 
given AMMonitor database that can be used to generate a 
cell app via ArcGIS Survey123 Connect. After running 
the function, the Excel file will be written to the user's 
working directory. 


This function should be used after updating the database 
with any new locations, people, equipment, user fields, etc. 
and the .xlsx file in Survey123 Connect should be 
updated accordingly. This will prevent data entry errors 
where a new location, person, etc. is not available for 
survey users. 

Please see the "mobile_apps" tutorial by running 
\code{learnr::run_tutorial("mobile_apps", package = "AMMonitor")}
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in a temporary directory
# (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# Set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# Create the Survey123 spreadsheet
filepath <- s123Create(
  con = conx, 
  reminders = c(
    'Is the case secured?',
    'Is the memory card secured?'),
  notes = c('Remember to label your SD card slip!'),
  disconnect = TRUE
)

# look at the filepath of the Excel file
filepath

# remove the demo
unlink(demo_fp)
  
}
}
