% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbInsertMetadata.R
\name{sbInsertMetadata}
\alias{sbInsertMetadata}
\title{Retrieve metadata about a ScienceBase release and insert it into the 
database}
\usage{
sbInsertMetadata(con, sbItemNum, xmlPath, fromRehydrate = TRUE, 
disconnect = FALSE )
}
\arguments{
\item{con}{A connection to the AMMonitor database}

\item{sbItemNum}{The ScienceBase item ID for an AMMonitor project data 
release on ScienceBase.}

\item{xmlPath}{The file path to the XML file containing metadata about the
ScienceBase release.}

\item{fromRehydrate}{TRUE or FALSE. Is the function called from sbRehydrate?
Default is TRUE}

\item{disconnect}{TRUE or FALSE, whether to disconnect from the database 
after the function is run. Default is FALSE}
}
\value{
This function does not return a value, but will add the metadata 
about a release
as a row in the sciencebase table in the AMMonitor database.
}
\description{
\code{sbInsertMetadata} retrieves metadata about a ScienceBase 
item and inserts it into an AMMonitor database.
}
\details{
#' For more information, please see the "sciencebase" learnr tutorial:
\code{learnr::run_tutorial(name = "sciencebase", package = "AMMonitor")}.
}
\examples{
\dontrun{
# --------------------------------------------
# for this example, we will assume that the demo database is a "master"
# --------------------------------------------

# create the mini demo project
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite")) 

# list the tables of the database
DBI::dbListTables(conx)

# let's assume the data were released on ScienceBase; have a look
browseURL("https://www.sciencebase.gov/catalog/item/654a576bd34ee4b6e05c24d6")

# the master database needs to be associated with this release

# note the master sciencebase table is empty 
DBI::dbReadTable(con, name = 'sciencebase')

# note the media files are not associated with the release (fk_sciencebaseid)
DBI::dbReadTable(con, name = 'media')

# now let's insert the release info into the "master" database  --------

# set Sciencebase item number
sbItemNum <- "654a576bd34ee4b6e05c24d6"

# download the release xml file  
fp <- sbtools::item_file_download(
  sb_id = sbItemNum,
  names = "Middle Earth Wildlife Study Volume 1 (2023 - 2023).xml",
  destinations = "Middle Earth Wildlife Study Volume 1 (2023 - 2023).xml")
  

# insert the metadata to the database
sbInsertMetadata(
  con = con,
  sbItemNum = "654a576bd34ee4b6e05c24d6",
  xmlPath = fp,
  fromRehydrate = FALSE,
  disconnect = FALSE
 )

# look at the sciencebase table again: note the primary key = 1
DBI::dbReadTable(con, name = 'sciencebase')

# confirm the media files are now associated with release 1
DBI::dbReadTable(con, name = 'media')

}
}
