% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbRehydrate.R
\name{sbRehydrate}
\alias{sbRehydrate}
\title{Recreate any AMMonitor project from a ScienceBase release}
\usage{
sbRehydrate(ammPath = getwd(), sbItemNum, projectName = NULL)
}
\arguments{
\item{ammPath}{A path to the directory which will house the AMMonitor project.
Default is getwd().}

\item{sbItemNum}{The item ID for an AMMonitor project data release on ScienceBase.}

\item{projectName}{A name for the newly created AMMonitor project. If not
provided, the title of the release will be used.}
}
\value{
Returns the filepath to the rehydrated AMMonitor project.
}
\description{
\code{sbRehydrate} recreates an AMMonitor project using files 
included in
a data release of an AMMonitor project on ScienceBase.
}
\details{
The sbRehydrate() function can be used to "rehydrate" data releases
published on the AMMonitor ScienceBase community.  This community is found 
at https://www.sciencebase.gov/catalog/item/6188c0c4d34ec04fc9c4f7a4. Each 
project is an AMMonitor project that has been released following strict
guidelines for releases on this USGS repository.  The sbRehydrate() function
will download all files associated with a release, and then will create a 
new AMMonitor project and database, ready for analysis.  The media files
associated with a release are linked as "child" items and should be downloaded
separately (they are very large zip files).

For more information, please see the "sciencebase" learnr tutorial:
\code{learnr::run_tutorial(name = "sciencebase", package = "AMMonitor")}.
}
\examples{
\dontrun{
# look at a release to be rehydrated 
browseURL("https://www.sciencebase.gov/catalog/item/654a576bd34ee4b6e05c24d6")

# create a rehydrated AMMonitor project (to be deleted)
project_fp <- sbRehydrate(
  ammPath = getwd(), 
  sbItemNum = "654a576bd34ee4b6e05c24d6",
  projectName = "MiddleEarth"
)

# look at the the filepath of the rehydrated project 
project_fp

# look at the folders within an AMMonitor project
list.files(project_fp, recursive = TRUE)

# look at the media directories - they are empty! You 
# need to download the media files manually
list.files(file.path(project_fp, "photos"), recursive = FALSE)
list.files(file.path(project_fp, "recordings"), recursive = FALSE)

# see who the default user is for the database
readLines(file.path(project_fp, "settings", "default_user.txt"))

# look at the sample database in the database directory
dbname <- list.files(file.path(project_fp, "database"))
dbname

# to work with the database, set a connection
conx <- dbSetCon(file.path(project_fp, "database", dbname))

# list the tables in the AMMonitor database
DBI::dbListTables(conx)

# the release information is stored within the "sciencebase" table
DBI::dbReadTable(conx, "sciencebase")

# note the media files are now associated with the release via fk_sciencebaseid
DBI::dbReadTable(conx, "media")

# run the database checks to ensure that rehydrated database has no issues
dbCheckup(conx)
dbCheckData(conx)
dbCheckCore(conx)

# disconnect from the database when finished
DBI::dbDisconnect(conx)

# to work with the Shiny app, just point to the AMMonitor project directory
# the app will set the database connection and disconnect
launchApp(amm_project_path = project_fp)

# remove the rehydrated project 
unlink(project_fp, recursive = TRUE, force = TRUE)

}
}
