% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoresGetFeatures.R
\name{scoresGetFeatures}
\alias{scoresGetFeatures}
\title{Get the spectrogram features of each detection by a particular template.}
\usage{
scoresGetFeatures(
  con,
  templateName,
  recordingRootPath = "/recordings/",
  showProgress = FALSE,
  disconnect = FALSE
)
}
\arguments{
\item{con}{A connection to a SQLite or Postgres AMMonitor database.}

\item{templateName}{A character vector specifying which template (by name) to get the outputs for.}

\item{recordingRootPath}{A character string indicating the path to the directory 
containing the recordings, can be local or a URL. Defaults to the local recordings
folder of an AMMonitor project}

\item{showProgress}{Default = FALSE. If TRUE, progress will be reported during 
the spectrogram retrieval.}

\item{disconnect}{Default = FALSE. If TRUE, the connection to the database will be closed
upon exit from the function}
}
\value{
dataframe of verifications and the spectrogram values as columns
}
\description{
\code{scoresGetFeatures} obtains the spectrogram for each verified 
detection by a particular template and associates them with the model verifications
so that they can be used to train classifier models. Only outputs where there 
is a consensus in the verifications (everyone who verified agreed that the output
is a true or false positive) will be obtained. While this function is designed 
for use with template outputs, it can also be used with other models that produce
outputs with identically sized bounding boxes.
}
\examples{
\dontrun{
# create a demo AMMonitor project (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# look at the models table
# notices that models 1:4 are templates created in AMMonitor
DBI::dbReadTable(conn = conx, name = "models")

# look at the modeloutputs table
# Note that model 3, oven_ct, has three outputs, 7:9
DBI::dbReadTable(conn = conx, name = "modeloutputs")

# look at the modelverifications table
# note that outputs 8 and 9 have been verified
DBI::dbReadTable(conn = conx, name = "modelverifications")

# retrieve the spectrograms for verified outputs of model 3, oven_ct
spectrograms <- scoresGetFeatures(
  con = conx, 
  templateName = "oven_ct", 
  recordingRootPath = paste0(demo_fp, "/recordings/"),
  disconnect = TRUE)
  
# unlink the demo
unlink(demo_fp)
}
}
