% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenNewVisitMetadata.R
\name{screenNewVisitMetadata}
\alias{screenNewVisitMetadata}
\title{Check new visits for potential errors before they are added to an
AMMonitor database}
\usage{
screenNewVisitMetadata(con, visitMetadata = NA, disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database}

\item{visitMetadata}{A data frame containing the metadata of the new visit.
Columns should be named to match the AMMonitor "visits" table.}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed on
exit? Default is FALSE.}
}
\value{
A list containing the status of the survey response (1 or 0) and a
data frame of any issues found with possible solutions.
}
\description{
Runs a set of checks on new visit metadata and compares it to
existing database records to check for errors and offers possible solutions
to errors found
}
\details{
This function runs a set of basic checks on the metadata of a new visit to
be added to the database to ensure there aren't any logic errors or other
issues. These checks include:

\itemize{
 \item Are all of the required fields for a new visit (fk_locationid,
 fk_equipmentid, fk_personid, and visit_type) provided?
 \item Does the visit already exist in the database?
 \item If the visit type is "set," is the equipment already currently deployed?
 \item If the visit type is "check" or "pull," is there a matching "set" or
 "check" visit in the database?
 }

The function will then return a list containing a status and a data frame of
warnings (if any). A status of 1 indicates that the visit could be added to
the database; a status of 0 indicates a major error that would prevent the
visit from being added. The warnings data frame includes the warning message,
a possible solution, and a severity level. It is recommended that all errors
are resolved before the visit is added to the database.
}
\examples{

\dontrun{

# Create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Set a connection to the demo database
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# Create a data frame with some example visit metadata
visit <- data.frame(
  pk_visitID = NA,
  fk_personid = "fbaggins",
  fk_locationid = "locationA",
  fk_equipmentid = "camera1",
  visit_type = "set",
  visit_date = "2024-01-01",
  visit_time = "12:00:00"
  )

# Screen the new visit's metadata
visit_check <- screenNewVisitMetadata(
  con = conx,
  visitMetadata = visit,
  disconnect = FALSE)

# Check the status of the check
visit_check$status

# Check the provided warnings
visit_check$warnings

# Run on a visit with missing data
visit <- data.frame(
  pk_visitID = NA,
  fk_personid = "fbaggins",
  fk_locationid = NA,
  fk_equipmentid = "camera1",
  visit_type = "set",
  visit_date = "2024-01-01",
  visit_time = "12:00:00"
  )

visit_check <- screenNewVisitMetadata(
  con = conx,
  visitMetadata = visit,
  disconnect = FALSE)

# Check the status of the check
visit_check$status

# Check the provided warnings
visit_check$warnings

}

}
