% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbInsertMetadata.R
\name{searchText}
\alias{searchText}
\title{Search text for every instance of a pattern in a character vector, and return only those instances}
\usage{
searchText(pattern, replacement, x)
}
\arguments{
\item{pattern}{Regex pattern to be passed to grepl and gsub}

\item{replacement}{Replacement for pattern, passed to gsub}

\item{x}{Character vector to search}
}
\value{
A character vector of the same length as x or less, with only the instances of the pattern
}
\description{
Searches a vector of character strings for a pattern, and changes them
to empty strings if the pattern isn't found.
}
