% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{stations}
\alias{stations}
\title{Sample weather stations for ammCreateMiniDemo}
\format{
A data frame with 14 rows and 18 variables
}
\usage{
data(stations)
}
\description{
Sample weather stations for ammCreateMiniDemo.
These stations were obtained with the `locationsGetStations()`
function for the demo.  The stations can be inserted into the 
locations table for illustration.  
\itemize{
  \item pk_locationid Unique name of weather station, no spaces.
  \item spatial_geometry Type of spatial geometry (point, line, polygon).  
  Acceptable entries are stored in the spatial_geometry list. 
  \item location_type Type of location (e.g., study area, monitoring station, 
  weather station, town). Here, "weather_station".
  \item fk_spatialid Foreign key to a spatial layer if the location is 
  represented by that layer (appropriate for study areas).
  \item lat Latitude (required for point geometries).
  \item long Longitude (required for point geometries).
  \item datum Datum - the model that represents the Earth. E.g. WGS84.  
  Acceptable entries are stored in the datum list (which can be added to).
  \item description Description of the location.
  \item x X coordinate if longitude is not provided for point locations.
  \item y Y coordinate if latitude is not provided for point locations.
  \item epsg EPSG code identifying the spatial attributes of the x and y 
  columns for point locations.  See https://epsg.org/home.html for 
  additional information.
  \item tz Olsen formatted timezone. See https://en.wikipedia.org/wiki
  /List_of_tz_database_time_zones.
  \item sensitive Is the location sensitive, such that it should not 
  included in data releases?
  \item long_min Minimum longitude for a bounding box, to be used for 
  obscuring the exact location of a monitoring station for publicly released 
  data.
  \item long_max Maximum longitude for a bounding box, to be used for 
  obscuring the exact location of a monitoring station for publicly released 
  data.
  \item lat_min Minimum latitude for a bounding box, to be used for obscuring 
  the exact location of a monitoring station for publicly released data.
  \item lat_max Maximum latitude for a bounding box, to be used for obscuring 
  the exact location of a monitoring station for publicly released data.
  \item location_status Status of location (e.g., active, retired, planned). 
   Acceptable entries are stored in the location_status list (which can be 
   modified but not deleted).
}
}
\keyword{datasets}
