% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{taxa}
\alias{taxa}
\title{Sample taxa table for ammCreateMiniDemo function}
\format{
A data frame with 15 rows and 16 variables
}
\usage{
data(taxa)
}
\description{
Sample taxa table for ammCreateMiniDemo function
\itemize{
  \item pk_taxonid Default is NA (will be given number when imported)
  \item common_name References pk_mediaid in media table
  \item tsn xx
  \item taxon_rank xx.
  \item rank_subspecies xx.
  \item rank_species xx.
  \item rank_genus x
  \item rank_family x
  \item rank_order x
  \item rank_class x
  \item rank_phylum x
  \item rank_kingdom x
  \item notes x
  \item sensitive x
  \item taxa_image_url x
  \item taxa_url x
}
}
\keyword{datasets}
