% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaAdd.R
\name{taxaAdd}
\alias{taxaAdd}
\title{Add a taxon to the taxa table, based on taxonomic serial number (TSN).}
\usage{
taxaAdd(con, tsns, common_names, pk_taxonids = NA, overwrite = FALSE, 
  disconnect = FALSE)
}
\arguments{
\item{con}{An existing database connection.}

\item{tsns}{Vector of taxonomic serial numbers (TSN's). See itis.gov}

\item{common_names}{Vector of preferred common names for the corresponding
TSN's}

\item{pk_taxonids}{Vector of primary key values for the corresponding taxon
(defaults to common name if not provided)}

\item{overwrite}{TRUE or FALSE, whether to overwrite existing taxon. Default
is FALSE}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed 
on exit? Default is FALSE}
}
\value{
A dataframe indicating whether each TSN was added, and if not, the
reason why not.
}
\description{
Given parallel vectors containing TSN, preferred common name, 
and short-hand text for the taxon (to be used as the primary key), this 
function uses the ritis R package to get the relevant taxonomic tree for the 
specified taxon and add to the database taxa table.
}
\details{
Please see the "taxa" learnr tutorial for more information:

\code{learnr::run_tutorial(name = "taxa", package = "AMMonitor")}
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in a temporary directory
# (to be deleted)

# run the function and capture the connection
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# retrieve the demo taxa table
taxa <- DBI::dbReadTable(conx, name = "taxa")

# look at the taxa table; note the column "tsn" (taxonomic serial number)
taxa

# use the package ritis to find a TSN for Red-eyed Vireo
ritis::search_common(x = "red-eyed vireo")

# add the Red-eyed Vireo to the taxa table
taxaAdd(
  con = conx, 
  tsns = 179021,
  common_names = "Red-eyed Vireo",
  pk_taxonids = "revi",
  overwrite = FALSE,
  disconnect = FALSE
)

# query the database for the new record
DBI::dbGetQuery(conx, statement = "SELECT * FROM taxa WHERE pk_taxonid = 'revi';")

# disconnect from the database when finished
DBI::dbDisconnect(conx)

# remove the demo
unlink(demo_fp)
}
}
\references{
For more information about Taxonomic Serial Numbers:
\href{https://www.itis.gov/}{https://www.itis.gov/}
}
