% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{temporals}
\alias{temporals}
\title{Sample weather data for ammCreateMiniDemo function}
\format{
A data frame with 4226 rows and 7 variables
}
\usage{
data(stations)
}
\description{
Sample weather data for ammCreateMiniDemo function.
These data were obtained with the `temporalsGet()`
function for the demo.  
\itemize{
  \item pk_temporalid Autonumber identifying a given temporal record.
  \item fk_locationid Maps to a given pk_locationid from the locations table.
  \item temporal_type Type of temporal record; e.g, historic or forecast.  
  Acceptable entries are stored in the list named temporal_type.
  \item temporal_date Date of the the temporal record.
  \item temporal_time Time of the the temporal record.
  \item k_temporallistitemid Maps to a given pk_temporallistitem in the 
  temporallistitems table.  Typically provides the name of a temporal variable.
  \item temporal_value_num Numeric value associated with the referenced 
  fk_temporallistitemid (if applicable).
}
}
\keyword{datasets}
