% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalsGet.R
\name{temporalsGet}
\alias{temporalsGet}
\title{Retrieves and formats temporal data}
\usage{
temporalsGet(amm_fp, con, startDate, endDate, token,
   temporalSource = "noaa", stationCodes = NULL, type = "historical",
   keys = c("TMAX", "TMIN", "PRCP", "SNOW", "SNWD"), dbInsert = FALSE,
   disconnect = FALSE)
}
\arguments{
\item{amm_fp}{Filepath to the AMMonitor project directory.  Default is getwd().}

\item{con}{A connection to the database}

\item{startDate}{Temporal data retrieved includes this date, must be a string 
in yyyy-mm-dd format}

\item{endDate}{Temporal data retrieved includes this date, must be a string in 
yyyy-mm-dd format}

\item{token}{Token obtained from 
\url{https://www.ncdc.noaa.gov/cdo-web/token}.  Options are "settings" or
"prompt".  Default value is "prompt", which prompts a token entry in the 
R console. If the token is stored
in the "settings" directory of your AMMonitor project as a text file,
enter "settings" for this argument. See details below.}

\item{temporalSource}{Source of temporal data, currently only supports "noaa"}

\item{stationCodes}{List of noaa weather station codes to retrieve temporal 
data for, if null, will retrieve stations from the database}

\item{type}{Type of temporal data to retrieve, either "historical" or "forecast". 
NOAA only provides historical data}

\item{keys}{Vector of variable names indicating which types of data to retrieve.
e.g. "TMAX" for maximum temperature, "TMIN" for minimum temperature, "PRCP" 
for precipitation, etc}

\item{dbInsert}{Whether or not to write these locations to the database}

\item{disconnect}{TRUE or FALSE. Should the database connection be severed
on exit? Default is FALSE}
}
\description{
This function retrieves temporal data from the National Climate 
Data Center (NCDC) or other specified sources.
}
\details{
This function can retrieve daily weather data from the
Global Historical Climatology Network (GHCN) using the REST API provided by NOAA.
Due to restrictions in the API, data can only be retrieved from a maximum time 
frame of one (1) year,
though we recommend breaking up time frames into shorter periods. There are no 
restrictions
on the number of stations that can be entered at once. Weather data from stations
can be retrieved by their station codes only. Please run \code{locationsGetStations()}
to retrieve station codes, and insert them in the database if planning to 
insert weather data into the database.

Different types of weather data can also be specified through the weatherType
argument by their abbreviated names (e.g. "TMAX" for maximum temperature,
"TMIN" for minimum temperature, etc.). By default, maximum temperature,
 minimum
temperature, precipitation, snowfall, and snow depth data are retrieved.
A complete list of weather data types, their abbreviations, and units can be 
found here:

\url{https://docs.opendata.aws/noaa-ghcn-pds/readme.html}

Please see the "temporals" tutorial for more information:

\code{learnr::run_tutorial(name = "temporals", package = "AMMonitor")}
}
\examples{
\dontrun{

# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the locations table; we are going to add weather stations
DBI::dbReadTable(conx, "locations")
  
# get the weather stations with dbInsert = TRUE
stations <- locationsGetStations(
  amm_fp = demo_fp,
  conx,
  noaa_token = "prompt",
  startDate = NULL,
  endDate = NULL,
  bbox =  c(43.15, -73, 43.5, -72.6),  # minlat, minlong, maxlat, maxlong
  dbInsert = TRUE,
  disconnect = FALSE)
 
# look at the stations
str(stations)
 
# get weather associated with stations
temporal_data <- temporalsGet(
  amm_fp  = demo_fp,
  con = conx, 
  startDate = "2023-01-01",
  endDate = "2023-12-31",
  token = "prompt",
  temporalSource = "noaa",
  stationCodes = stations$pk_locationid,
  type = "historical",
  keys = c("TMAX", "TMIN", "PRCP", "SNOW", "SNWD"),
  dbInsert = FALSE,
  disconnect = FALSE
 )
 
# look at the retrieved data
head(temporal_data)
 
# insert to database if appropriate
DBI::dbAppendTable(conx, name = "temporals", value = temporal_data)
 
# look at the data in the temporals table
DBI::dbReadTable(conx, name = "temporals")
 
# disconnect from the database when finished
DBI::dbDisconnect(conx)
 
# remove the demo AMMonitor file structure
unlink(demo_fp, recursive = TRUE)

}
}
\seealso{
Other data: 
\code{\link{locationsGetStations}()}
}
\concept{data}
\keyword{data}
