% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyExcelImport.R
\name{verifyExcelImport}
\alias{verifyExcelImport}
\title{Import annotation and modeloutput verifications from an Excel 
spreadsheet}
\usage{
verifyExcelImport(con, excel_file, verified_what, person_id, 
disconnect = FALSE)
}
\arguments{
\item{con}{An open connection to an AMMonitor database.}

\item{excel_file}{The file path to the excel file containing the 
verifications to be imported.}

\item{verified_what}{Were annotations or model outputs verified? Options are 
"annotations" or "modeloutputs".}

\item{person_id}{The pk_personid of the person who completed the 
verifications, as stored in the "people" table of the database}

\item{disconnect}{TRUE or FALSE. Should the connection to the database be
severed on exit? Default is FALSE.}
}
\description{
Import verifications from an Excel spreadsheet created with
verifyExcelCreate to an AMMonitor database
}
\details{
This function takes a verifications spreadsheet, created using 
verifyExcelCreate(), and imports all verifications into the AMMonitor 
database. Note that all verifications in the Excel file must be input as TRUE
or FALSE for this function to work properly. For more details on creating
verification spreadsheets, see the verifyExcelCreate() helpfile.

The user must specify what was verified, "annotations" or "modeloutputs", as 
well as who completed the verifications using the pk_personid listed in the
database's "people" table. These details will be used to add all 
verifications to the proper table. If any outputs were not verified and left
blank, no new record will be added to the verifications table for that 
output.

For more details on annotations and verifications, please check the learnr 
"annotations" tutorial using the following code:
\code{learnr::run_tutorial(name = "annotations", package = "AMMonitor")}
}
\examples{
\dontrun{

# Create an AMMonitor demo database, to be deleted later
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# Set a connection to the database
conx <- dbSetCon(paste0(demo_fp, "/database/demo.sqlite"))

# Get the media IDs to create a verification spreadsheet
media_ids <- DBI::dbGetQuery(
  conn = conx,
  statement = "SELECT pk_mediaid FROM media;"
  )[,,drop = TRUE]
  
# Set a file path for the output Excel sheet (including filename)
outfile_path <- paste0(demo_fp, "/validate_me.xlsx")

# Create the verifications spreadsheet
results <- verifyExcelCreate(
 con = conx, 
 media_ids = media_ids,
 verify_what = "modeloutputs",
 media_root_path = 
 "https://code.usgs.gov/vtcfwru/ammonitor/-/raw/AMMonitor2.2/inst/extdata/demoAMM",
 outfile_path =  outfile_path,
 disconnect = FALSE)

## NOTE: Open up your new verification spreadsheet and add some validations
## to the "verified" column using TRUE and FALSE

# Import the verifications from the output file as gandalf
verifyExcelImport(
 con = conx,
 excel_file = outfile_path,
 verified_what = "modeloutputs",
 person_id = "gandalf",
 disconnect = FALSE)

# Look at the model verifications table to see the added verifications
View(DBI::dbReadTable(conx, "modelverifications"))

}

}
\seealso{
Other verifications: 
\code{\link{verifyExcelCreate}()}
}
\concept{verifications}
