% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qry.R
\name{qryModelOutputsMedia}
\alias{qryModelOutputsMedia}
\title{Filter a list of media with model outputs based on many criteria}
\usage{
qryModelOutputsMedia(
  con, 
  disconnect = FALSE, 
  locationID = 'all', 
  dateRange = list(c('1900-01-01', as.character(Sys.Date()))), 
  visitID = NULL, 
  taxonID = 'all', 
  excludeAnnoVerified = 'NA',
  selectedUser = NA, 
  model = "all", 
  confValue = 0, 
  lessThan = FALSE, 
  newOnly = FALSE, 
  mediaType = "photo",
  limit = Inf,
  offset = 0
)
}
\arguments{
\item{con}{An open SQLite database connection}

\item{disconnect}{TRUE or FALSE. Should the database connection be closed on 
exit?  Default is FALSE.}

\item{locationID}{Specify a location (pk_locationid) from the locations
table. Default is "all".}

\item{dateRange}{Specify a date range. Default is "1900-01-01" to current
date.}

\item{visitID}{Specify a specific visit (pk_visitid) from the visits table.
Default is NULL.}

\item{taxonID}{Specify a specific taxon (pk_taxonid) from the taxa table. 
Default is "all".}

\item{excludeAnnoVerified}{Specify when to exclude based on verification 
status. Default is "NA".}

\item{confValue}{Specify a confidence level to filter modeloutputs 
(model_value).  Default is 0.}

\item{lessThan}{TRUE or FALSE. Specify if outputs should be less than or 
greater than confValue. Default is FALSE.}

\item{selectedUser}{The fk_personid who verified modeloutputs
(pk_personid from the people table). Default is NA.}

\item{model}{Which model to display (pk_modelid from the models table). 
Default is "all".}

\item{mediaType}{Type of media. Options are "photo" (default) or "audio".}

\item{newOnly}{TRUE or FALSE. Default is FALSE.}

\item{limit}{Maximum number of records to be returned}

\item{offset}{The number of records to offset the query by}
}
\value{
A dataframe with key columns from all photos that meet the given criteria
}
\description{
Filter a list of media based on many criteria, but only include
media with model outputs.
}
\examples{
\dontrun{
# create a demo AMMonitor project in a temporary directory (to be deleted)
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# look at the models table
DBI::dbReadTable(conx, name = "models")

# look at the modeloutputs table
DBI::dbReadTable(conx, name = "modeloutputs")

# look at the modelverifications table
DBI::dbReadTable(conx, name = "modelverifications")

# run the query to get media that have modeloutputs with "btnw"
qryModelOutputsMedia(conx, taxonID = 'btnw', mediaType = "audio")

# get media files that were tagged by MegaDetector (model 5)
qryModelOutputsMedia(conx,  mediaType = "photo", model = 5)

# get media photo files that have a high confidence score
qryModelOutputsMedia(conx,  mediaType = "photo", confValue = 0.9, disconnect = TRUE)

# unlink the demo
unlink(demo_fp)
}
}
\seealso{
Other query: 
\code{\link{qryCheckMediaFilePaths}()},
\code{\link{qryCheckTags}()},
\code{\link{qryEconfiguration}()},
\code{\link{qryEffort}()},
\code{\link{qryEquipModelSettings}()},
\code{\link{qryItems}()},
\code{\link{qryLocationCounts}()},
\code{\link{qryMconfiguration}()},
\code{\link{qryMedia}()},
\code{\link{qryMediaDateRange}()},
\code{\link{qryMediaLocations}()},
\code{\link{qryModelOutputs}()},
\code{\link{qryModelOutputsFull}()},
\code{\link{qryModelSettings}()},
\code{\link{qryModelVerificationConsensus}()},
\code{\link{qryPeopleActivity}()},
\code{\link{qryTags}()},
\code{\link{qryUnannotated}()},
\code{\link{qryVerifications}()},
\code{\link{qryVisitTable}()}
}
\concept{media}
\concept{modeloutputs}
\concept{models}
\concept{query}
