% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoresDetect.R
\name{scoresDetect}
\alias{scoresDetect}
\title{Detect events using \code{\link[monitoR:makeTemplate]{monitoR}}
templates}
\usage{
scoresDetect(
  con,
  recordingNames = "all",
  templateNames = "all",
  scoreThresholds = NA,
  recordingRootPath = "recordings/",
  ammlPath = "amml/",
  dbInsert = FALSE,
  showProgress = FALSE,
  corMethod = "pearson"
)
}
\arguments{
\item{con}{A connection to a SQLite or Postgres AMMonitor database.}

\item{recordingNames}{Default = 'all'. An optional character vector specifying
by name which recordings should be processed. Default is 'all'.}

\item{templateNames}{Default = 'all'. An optional character vector specifying
which templates to use for searching recordings.}

\item{scoreThresholds}{Optional named numeric vector of score thresholds to
use for binary point matching or spectrogram cross correlation templates,
ordered according to inputs to the templateID or listID argument. If using
\code{templateID = 'all'}, please check the current ordering of templates in
your database to ensure you are providing the correct score thresholds to the
templateNames you intend. If no value is entered for scoreThresholds, all
template score thresholds will be automatically extracted from the internals
of the monitoR templateList object -- please check to ensure these are the
score thresholds you actually want to use.}

\item{recordingRootPath}{A character string indicating the path to the directory
containing the recordings, can be local or a URL.}

\item{ammlPath}{A character string indicating the path to the directory where
AMModel Libraries are stored.}

\item{dbInsert}{Default = FALSE. Logical flag for whether to insert records
generated by this function into the database. If TRUE, events are added to the
database. If FALSE, events are returned as a data.table for examination by the
user, but not added to the database. Setting dbInsert = TRUE will also remove
duplicate records before inserting into the database.}

\item{showProgress}{Default = FALSE. If TRUE, progress will be reported during
the score calculations.}

\item{corMethod}{Default = 'pearson'. For spectrogram cross correlation templates,
the method used to calculate correlation coefficients (see \code{?cor}).}
}
\value{
Data.table of scores
}
\description{
\code{scoresDetect} calls monitoR's \code{templateMatching()}
function to run spectrogram cross correlation or binary point matching templates
in search of target signals within recordings. While running template matching,
\code{scoresDetect} also extracts acoustic features associated with each
detected event (see the Details section). To indicate which recordings should
be run in \code{scoresDetect}, supply the recordingNames. To indicate which
templates should be run against recordings, provide the name(s) of templates
as stored in the models table of the AMMonitor database, and further supply
the path to the AMModels library that stores the RDS template library.
}
\details{
The recording root path argument can be NA.
}
\examples{
\dontrun{

# create a demo AMMonitor file structure in a temporary directory
# (to be deleted)

# run the function and capture the connection
demo_fp <- ammCreateMiniDemo(filepath = tempdir())

# look at the demo_fp
demo_fp

# look at the sample recordings in the recordings directory
list.files(paste0(demo_fp, "/recordings"), recursive = FALSE)

# look at the sample ammodels library that stores templates in the amml directory
list.files(paste0(demo_fp, "/ammls"), recursive = FALSE)

# read in the "templates" model library; note there are 4 templates
templates <- readRDS(paste0(demo_fp, "/ammls/templates.RDS"))
templates

# to work with the database, set a connection
conx <- dbSetCon(file.path(demo_fp, "database", "demo.sqlite"))

# Let's look at the modeloutputs table
DBI::dbReadTable(conx, name = 'modeloutputs')

# clear out the pre-populated modeloutputs table
DBI::dbExecute(conx, statement = "DELETE FROM modeloutputs;")

# run scoresDetect to match the 4 templates against 1.wav
scores <- scoresDetect(
  con = conx,
  recordingNames = "1.wav",
  recordingRootPath = paste0(demo_fp, "/recordings"),
  ammlPath = paste0(demo_fp, "/ammls"),
  dbInsert = FALSE
 )

 # look at the scores
 scores

# rerun the function is dbInsert is TRUE
scoresDetect(
  con = conx,
  recordingNames = "1.wav",
  recordingRootPath = paste0(demo_fp, "/recordings"),
  ammlPath = paste0(demo_fp, "/ammls"),
  dbInsert = TRUE
 )
 
 # rerun the function with scoreThresholds specified
 
 newThresholds <- c(.55, 16, .55, 14)
 names(newThresholds) <-  names(templates@models)


 scores_newThreshold <- scoresDetect(
  con = conx,
  recordingNames = "1.wav",
  scoreThresholds = newThresholds,
  recordingRootPath = paste0(demo_fp, "/recordings"),
  ammlPath = paste0(demo_fp, "/ammls"),
  dbInsert = FALSE
 )

# check the modeloutputs table
DBI::dbReadTable(conx, name = 'modeloutputs')

}
}
